refactor(codegen): generate access/error/proxy once
The access and proxy templates already gate minio operations; generate
them as single files like s3_trait.rs and drop the _minio variants.
The error module depends on the static error-code table only (the minio
patch adds no error codes), so its files were identical and it is also
generated once. Consumer modules drop the cfg_if!/dual-mod declarations
in favor of a plain mod generated.

The proxy template now emits #[cfg(feature = "minio")] on the
MinIO-only forwarding methods; s3_trait/access emit doc comments before
the cfg gate, matching the single-file layout.
