+++ codegen/src/v1/access.rs
@@ -49,13 +49,12 @@ pub fn codegen(ops: &Operations) {
         let method_name = op.name.to_snake_case();
         let input = &op.input;
 
-        if op.is_minio {
-            g!("#[cfg(feature = \"minio\")]");
-        }
-
         g!("/// Checks whether the {} request has accesses to the resources.", op.name);
         g!("/// ");
         g!("/// This method returns `Ok(())` by default.");
+        if op.is_minio {
+            g!("#[cfg(feature = \"minio\")]");
+        }
         g!("async fn {method_name}(&self, _req: &mut S3Request<{input}>) -> S3Result<()> {{");
         g!("Ok(())");
         g!("}}");
