+++ crates/s3s-aws/src/proxy/generated.rs
@@ -1828,6 +1828,15 @@ impl S3 for Proxy {
     }
 
     #[tracing::instrument(skip(self, req))]
+    #[cfg(feature = "minio")]
+    async fn listen_bucket_notification(
+        &self,
+        req: S3Request<s3s::dto::ListenBucketNotificationInput>,
+    ) -> S3Result<S3Response<s3s::dto::ListenBucketNotificationOutput>> {
+        super::listen::listen_bucket_notification(&self.minio, req).await
+    }
+
+    #[tracing::instrument(skip(self, req))]
     async fn put_bucket_abac(
         &self,
         req: S3Request<s3s::dto::PutBucketAbacInput>,
