+++ crates/s3s-aws/src/proxy/mod.rs
@@ -1,10 +1,6 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2023-2026 The s3s Authors
 
-#[cfg(feature = "minio")]
-mod generated_minio;
-
-#[cfg(not(feature = "minio"))]
 mod generated;
 
 #[cfg(feature = "minio")]
