+++ crates/s3s/src/access/mod.rs
@@ -96,14 +96,7 @@
 //! let service = builder.build();
 //! ```
 
-cfg_if::cfg_if! {
-    if #[cfg(feature = "minio")] {
-        mod generated_minio;
-        use self::generated_minio as generated;
-    } else {
-        mod generated;
-    }
-}
+mod generated;
 
 pub use self::generated::S3Access;
