--- crates/s3s/src/error/generated_minio.rs
@@ -1,3914 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-// SPDX-FileCopyrightText: 2023-2026 The s3s Authors
-
-//! Auto generated by `s3s_codegen::v1::error::codegen`
-
-#![allow(clippy::doc_markdown)]
-#![allow(clippy::too_many_lines)]
-#![allow(clippy::unreadable_literal)]
-
-use bytestring::ByteString;
-use hyper::StatusCode;
-
-// AccessControlListNotSupported
-// AccessDenied
-// AccessPointAlreadyOwnedByYou
-// AccountProblem
-// AllAccessDisabled
-// AmbiguousFieldName
-// AmbiguousGrantByEmailAddress
-// AuthorizationHeaderMalformed
-// AuthorizationQueryParametersError
-// BadDigest
-// BucketAlreadyExists
-// BucketAlreadyOwnedByYou
-// BucketHasAccessPointsAttached
-// BucketNotEmpty
-// Busy
-// CSVEscapingRecordDelimiter
-// CSVParsingError
-// CSVUnescapedQuote
-// CastFailed
-// ClientTokenConflict
-// ColumnTooLong
-// ConditionalRequestConflict
-// ConnectionClosedByRequester
-// CredentialsNotSupported
-// CrossLocationLoggingProhibited
-// DeviceNotActiveError
-// EmptyRequestBody
-// EndpointNotFound
-// EntityTooLarge
-// EntityTooSmall
-// EvaluatorBindingDoesNotExist
-// EvaluatorInvalidArguments
-// EvaluatorInvalidTimestampFormatPattern
-// EvaluatorInvalidTimestampFormatPatternSymbol
-// EvaluatorInvalidTimestampFormatPatternSymbolForParsing
-// EvaluatorInvalidTimestampFormatPatternToken
-// EvaluatorLikePatternInvalidEscapeSequence
-// EvaluatorNegativeLimit
-// EvaluatorTimestampFormatPatternDuplicateFields
-// EvaluatorTimestampFormatPatternHourClockAmPmMismatch
-// EvaluatorUnterminatedTimestampFormatPatternToken
-// ExpiredToken
-// ExpressionTooLong
-// ExternalEvalException
-// IllegalLocationConstraintException
-// IllegalSqlFunctionArgument
-// IllegalVersioningConfigurationException
-// IncompleteBody
-// IncorrectEndpoint
-// IncorrectNumberOfFilesInPostRequest
-// IncorrectSqlFunctionArgumentType
-// InlineDataTooLarge
-// IntegerOverflow
-// InternalError
-// InvalidAccessKeyId
-// InvalidAccessPoint
-// InvalidAccessPointAliasError
-// InvalidAddressingHeader
-// InvalidArgument
-// InvalidBucketAclWithObjectOwnership
-// InvalidBucketName
-// InvalidBucketOwnerAWSAccountID
-// InvalidBucketState
-// InvalidCast
-// InvalidColumnIndex
-// InvalidCompressionFormat
-// InvalidDataSource
-// InvalidDataType
-// InvalidDigest
-// InvalidEncryptionAlgorithmError
-// InvalidExpressionType
-// InvalidFileHeaderInfo
-// InvalidHostHeader
-// InvalidHttpMethod
-// InvalidJsonType
-// InvalidKeyPath
-// InvalidLocationConstraint
-// InvalidObjectState
-// InvalidPart
-// InvalidPartOrder
-// InvalidPayer
-// InvalidPolicyDocument
-// InvalidQuoteFields
-// InvalidRange
-// InvalidRegion
-// InvalidRequest
-// InvalidRequestParameter
-// InvalidSOAPRequest
-// InvalidScanRange
-// InvalidSecurity
-// InvalidSessionException
-// InvalidSignature
-// InvalidStorageClass
-// InvalidTableAlias
-// InvalidTag
-// InvalidTargetBucketForLogging
-// InvalidTextEncoding
-// InvalidToken
-// InvalidURI
-// JSONParsingError
-// KeyTooLongError
-// LexerInvalidChar
-// LexerInvalidIONLiteral
-// LexerInvalidLiteral
-// LexerInvalidOperator
-// LikeInvalidInputs
-// MalformedACLError
-// MalformedPOSTRequest
-// MalformedPolicy
-// MalformedXML
-// MaxMessageLengthExceeded
-// MaxOperatorsExceeded
-// MaxPostPreDataLengthExceededError
-// MetadataTooLarge
-// MethodNotAllowed
-// MissingAttachment
-// MissingAuthenticationToken
-// MissingContentLength
-// MissingRequestBodyError
-// MissingRequiredParameter
-// MissingSecurityElement
-// MissingSecurityHeader
-// MultipleDataSourcesUnsupported
-// NoLoggingStatusForKey
-// NoSuchAccessPoint
-// NoSuchAsyncRequest
-// NoSuchBucket
-// NoSuchBucketPolicy
-// NoSuchCORSConfiguration
-// NoSuchKey
-// NoSuchLifecycleConfiguration
-// NoSuchMultiRegionAccessPoint
-// NoSuchObjectLockConfiguration
-// NoSuchResource
-// NoSuchTagSet
-// NoSuchUpload
-// NoSuchVersion
-// NoSuchWebsiteConfiguration
-// NoTransformationDefined
-// NotDeviceOwnerError
-// NotImplemented
-// NotModified
-// NotSignedUp
-// NumberFormatError
-// ObjectLockConfigurationNotFoundError
-// ObjectSerializationConflict
-// OperationAborted
-// OverMaxColumn
-// OverMaxParquetBlockSize
-// OverMaxRecordSize
-// OwnershipControlsNotFoundError
-// ParquetParsingError
-// ParquetUnsupportedCompressionCodec
-// ParseAsteriskIsNotAloneInSelectList
-// ParseCannotMixSqbAndWildcardInSelectList
-// ParseCastArity
-// ParseEmptySelect
-// ParseExpected2TokenTypes
-// ParseExpectedArgumentDelimiter
-// ParseExpectedDatePart
-// ParseExpectedExpression
-// ParseExpectedIdentForAlias
-// ParseExpectedIdentForAt
-// ParseExpectedIdentForGroupName
-// ParseExpectedKeyword
-// ParseExpectedLeftParenAfterCast
-// ParseExpectedLeftParenBuiltinFunctionCall
-// ParseExpectedLeftParenValueConstructor
-// ParseExpectedMember
-// ParseExpectedNumber
-// ParseExpectedRightParenBuiltinFunctionCall
-// ParseExpectedTokenType
-// ParseExpectedTypeName
-// ParseExpectedWhenClause
-// ParseInvalidContextForWildcardInSelectList
-// ParseInvalidPathComponent
-// ParseInvalidTypeParam
-// ParseMalformedJoin
-// ParseMissingIdentAfterAt
-// ParseNonUnaryAgregateFunctionCall
-// ParseSelectMissingFrom
-// ParseUnExpectedKeyword
-// ParseUnexpectedOperator
-// ParseUnexpectedTerm
-// ParseUnexpectedToken
-// ParseUnknownOperator
-// ParseUnsupportedAlias
-// ParseUnsupportedCallWithStar
-// ParseUnsupportedCase
-// ParseUnsupportedCaseClause
-// ParseUnsupportedLiteralsGroupBy
-// ParseUnsupportedSelect
-// ParseUnsupportedSyntax
-// ParseUnsupportedToken
-// PermanentRedirect
-// PermanentRedirectControlError
-// PreconditionFailed
-// Redirect
-// ReplicationConfigurationNotFoundError
-// RequestHeaderSectionTooLarge
-// RequestIsNotMultiPartContent
-// RequestTimeTooSkewed
-// RequestTimeout
-// RequestTorrentOfBucketError
-// ResponseInterrupted
-// RestoreAlreadyInProgress
-// ServerSideEncryptionConfigurationNotFoundError
-// ServiceUnavailable
-// SignatureDoesNotMatch
-// SlowDown
-// TagPolicyException
-// TemporaryRedirect
-// TokenCodeInvalidError
-// TokenRefreshRequired
-// TooManyAccessPoints
-// TooManyBuckets
-// TooManyMultiRegionAccessPointregionsError
-// TooManyMultiRegionAccessPoints
-// TooManyTags
-// TruncatedInput
-// UnauthorizedAccess
-// UnauthorizedAccessError
-// UnexpectedContent
-// UnexpectedIPError
-// UnrecognizedFormatException
-// UnresolvableGrantByEmailAddress
-// UnsupportedArgument
-// UnsupportedFunction
-// UnsupportedParquetType
-// UnsupportedRangeHeader
-// UnsupportedScanRangeInput
-// UnsupportedSignature
-// UnsupportedSqlOperation
-// UnsupportedSqlStructure
-// UnsupportedStorageClass
-// UnsupportedSyntax
-// UnsupportedTypeForQuerying
-// UserKeyMustBeSpecified
-// ValueParseFailure
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-#[non_exhaustive]
-pub enum S3ErrorCode {
-    /// The bucket does not allow ACLs.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    AccessControlListNotSupported,
-
-    /// Access Denied
-    ///
-    /// HTTP Status Code: 403 Forbidden
-    ///
-    AccessDenied,
-
-    /// An access point with an identical name already exists in your account.
-    ///
-    /// HTTP Status Code: 409 Conflict
-    ///
-    AccessPointAlreadyOwnedByYou,
-
-    /// There is a problem with your Amazon Web Services account that prevents the action from completing successfully. Contact Amazon Web Services Support for further assistance.
-    ///
-    /// HTTP Status Code: 403 Forbidden
-    ///
-    AccountProblem,
-
-    /// All access to this Amazon S3 resource has been disabled. Contact Amazon Web Services Support for further assistance.
-    ///
-    /// HTTP Status Code: 403 Forbidden
-    ///
-    AllAccessDisabled,
-
-    /// The field name matches to multiple fields in the file. Check the SQL expression and the file, and try again.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    AmbiguousFieldName,
-
-    /// The email address you provided is associated with more than one account.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    AmbiguousGrantByEmailAddress,
-
-    /// The authorization header you provided is invalid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    AuthorizationHeaderMalformed,
-
-    /// The authorization query parameters that you provided are not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    AuthorizationQueryParametersError,
-
-    /// The Content-MD5 you specified did not match what we received.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    BadDigest,
-
-    /// The requested bucket name is not available. The bucket namespace is shared by all users of the system. Please select a different name and try again.
-    ///
-    /// HTTP Status Code: 409 Conflict
-    ///
-    BucketAlreadyExists,
-
-    /// The bucket you tried to create already exists, and you own it. Amazon S3 returns this error in all Amazon Web Services Regions except in the North Virginia Region. For legacy compatibility, if you re-create an existing bucket that you already own in the North Virginia Region, Amazon S3 returns 200 OK and resets the bucket access control lists (ACLs).
-    ///
-    /// HTTP Status Code: 409 Conflict
-    ///
-    BucketAlreadyOwnedByYou,
-
-    /// The bucket you tried to delete has access points attached. Delete your access points before deleting your bucket.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    BucketHasAccessPointsAttached,
-
-    /// The bucket you tried to delete is not empty.
-    ///
-    /// HTTP Status Code: 409 Conflict
-    ///
-    BucketNotEmpty,
-
-    /// The service is unavailable. Try again later.
-    ///
-    /// HTTP Status Code: 503 Service Unavailable
-    ///
-    Busy,
-
-    /// A quoted record delimiter was found in the file. To allow quoted record delimiters, set AllowQuotedRecordDelimiter to 'TRUE'.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    CSVEscapingRecordDelimiter,
-
-    /// An error occurred while parsing the CSV file. Check the file and try again.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    CSVParsingError,
-
-    /// An unescaped quote was found while parsing the CSV file. To allow quoted record delimiters, set AllowQuotedRecordDelimiter to 'TRUE'.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    CSVUnescapedQuote,
-
-    /// An attempt to convert from one data type to another using CAST failed in the SQL expression.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    CastFailed,
-
-    /// Your Multi-Region Access Point idempotency token was already used for a different request.
-    ///
-    /// HTTP Status Code: 409 Conflict
-    ///
-    ClientTokenConflict,
-
-    /// The length of a column in the result is greater than maxCharsPerColumn of 1 MB.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ColumnTooLong,
-
-    /// A conflicting operation occurred. If using PutObject you can retry the request. If using multipart upload you should initiate another CreateMultipartUpload request and re-upload each part.
-    ///
-    /// HTTP Status Code: 409 Conflict
-    ///
-    ConditionalRequestConflict,
-
-    /// Returned to the original caller when an error is encountered while reading the WriteGetObjectResponse body.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ConnectionClosedByRequester,
-
-    /// This request does not support credentials.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    CredentialsNotSupported,
-
-    /// Cross-location logging not allowed. Buckets in one geographic location cannot log information to a bucket in another location.
-    ///
-    /// HTTP Status Code: 403 Forbidden
-    ///
-    CrossLocationLoggingProhibited,
-
-    /// The device is not currently active.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    DeviceNotActiveError,
-
-    /// The request body cannot be empty.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    EmptyRequestBody,
-
-    /// Direct requests to the correct endpoint.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    EndpointNotFound,
-
-    /// Your proposed upload exceeds the maximum allowed object size.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    EntityTooLarge,
-
-    /// Your proposed upload is smaller than the minimum allowed object size.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    EntityTooSmall,
-
-    /// A column name or a path provided does not exist in the SQL expression.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    EvaluatorBindingDoesNotExist,
-
-    /// There is an incorrect number of arguments in the function call in the SQL expression.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    EvaluatorInvalidArguments,
-
-    /// The timestamp format string in the SQL expression is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    EvaluatorInvalidTimestampFormatPattern,
-
-    /// The timestamp format pattern contains a symbol in the SQL expression that is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    EvaluatorInvalidTimestampFormatPatternSymbol,
-
-    /// The timestamp format pattern contains a valid format symbol that cannot be applied to timestamp parsing in the SQL expression.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    EvaluatorInvalidTimestampFormatPatternSymbolForParsing,
-
-    /// The timestamp format pattern contains a token in the SQL expression that is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    EvaluatorInvalidTimestampFormatPatternToken,
-
-    /// An argument given to the LIKE expression was not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    EvaluatorLikePatternInvalidEscapeSequence,
-
-    /// LIMIT must not be negative.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    EvaluatorNegativeLimit,
-
-    /// The timestamp format pattern contains multiple format specifiers representing the timestamp field in the SQL expression.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    EvaluatorTimestampFormatPatternDuplicateFields,
-
-    /// The timestamp format pattern contains a 12-hour hour of day format symbol but doesn't also contain an AM/PM field, or it contains a 24-hour hour of day format specifier and contains an AM/PM field in the SQL expression.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    EvaluatorTimestampFormatPatternHourClockAmPmMismatch,
-
-    /// The timestamp format pattern contains an unterminated token in the SQL expression.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    EvaluatorUnterminatedTimestampFormatPatternToken,
-
-    /// The provided token has expired.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ExpiredToken,
-
-    /// The SQL expression is too long. The maximum byte-length for an SQL expression is 256 KB.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ExpressionTooLong,
-
-    /// The query cannot be evaluated. Check the file and try again.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ExternalEvalException,
-
-    /// This error might occur for the following reasons:
-    ///
-    ///
-    /// You are trying to access a bucket from a different Region than where the bucket exists.
-    ///
-    /// You attempt to create a bucket with a location constraint that corresponds to a different region than the regional endpoint the request was sent to.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    IllegalLocationConstraintException,
-
-    /// An illegal argument was used in the SQL function.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    IllegalSqlFunctionArgument,
-
-    /// Indicates that the versioning configuration specified in the request is invalid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    IllegalVersioningConfigurationException,
-
-    /// You did not provide the number of bytes specified by the Content-Length HTTP header
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    IncompleteBody,
-
-    /// The specified bucket exists in another Region. Direct requests to the correct endpoint.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    IncorrectEndpoint,
-
-    /// POST requires exactly one file upload per request.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    IncorrectNumberOfFilesInPostRequest,
-
-    /// An incorrect argument type was specified in a function call in the SQL expression.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    IncorrectSqlFunctionArgumentType,
-
-    /// Inline data exceeds the maximum allowed size.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InlineDataTooLarge,
-
-    /// An integer overflow or underflow occurred in the SQL expression.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    IntegerOverflow,
-
-    /// We encountered an internal error. Please try again.
-    ///
-    /// HTTP Status Code: 500 Internal Server Error
-    ///
-    InternalError,
-
-    /// The Amazon Web Services access key ID you provided does not exist in our records.
-    ///
-    /// HTTP Status Code: 403 Forbidden
-    ///
-    InvalidAccessKeyId,
-
-    /// The specified access point name or account is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidAccessPoint,
-
-    /// The specified access point alias name is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidAccessPointAliasError,
-
-    /// You must specify the Anonymous role.
-    ///
-    InvalidAddressingHeader,
-
-    /// Invalid Argument
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidArgument,
-
-    /// Bucket cannot have ACLs set with ObjectOwnership's BucketOwnerEnforced setting.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidBucketAclWithObjectOwnership,
-
-    /// The specified bucket is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidBucketName,
-
-    /// The value of the expected bucket owner parameter must be an AWS account ID.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidBucketOwnerAWSAccountID,
-
-    /// The request is not valid with the current state of the bucket.
-    ///
-    /// HTTP Status Code: 409 Conflict
-    ///
-    InvalidBucketState,
-
-    /// An attempt to convert from one data type to another using CAST failed in the SQL expression.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidCast,
-
-    /// The column index in the SQL expression is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidColumnIndex,
-
-    /// The file is not in a supported compression format. Only GZIP and BZIP2 are supported.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidCompressionFormat,
-
-    /// The data source type is not valid. Only CSV, JSON, and Parquet are supported.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidDataSource,
-
-    /// The SQL expression contains a data type that is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidDataType,
-
-    /// The Content-MD5 you specified is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidDigest,
-
-    /// The encryption request you specified is not valid. The valid value is AES256.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidEncryptionAlgorithmError,
-
-    /// The ExpressionType value is not valid. Only SQL expressions are supported.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidExpressionType,
-
-    /// The FileHeaderInfo value is not valid. Only NONE, USE, and IGNORE are supported.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidFileHeaderInfo,
-
-    /// The host headers provided in the request used the incorrect style addressing.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidHostHeader,
-
-    /// The request is made using an unexpected HTTP method.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidHttpMethod,
-
-    /// The JsonType value is not valid. Only DOCUMENT and LINES are supported.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidJsonType,
-
-    /// The key path in the SQL expression is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidKeyPath,
-
-    /// The specified location constraint is not valid. For more information about Regions, see <a href="https://docs.aws.amazon.com/AmazonS3/latest/dev/UsingBucket.html#access-bucket-intro">How to Select a Region for Your Buckets</a>.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidLocationConstraint,
-
-    /// The action is not valid for the current state of the object.
-    ///
-    /// HTTP Status Code: 403 Forbidden
-    ///
-    InvalidObjectState,
-
-    /// One or more of the specified parts could not be found. The part might not have been uploaded, or the specified entity tag might not have matched the part's entity tag.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidPart,
-
-    /// The list of parts was not in ascending order. Parts list must be specified in order by part number.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidPartOrder,
-
-    /// All access to this object has been disabled. Please contact Amazon Web Services Support for further assistance.
-    ///
-    /// HTTP Status Code: 403 Forbidden
-    ///
-    InvalidPayer,
-
-    /// The content of the form does not meet the conditions specified in the policy document.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidPolicyDocument,
-
-    /// The QuoteFields value is not valid. Only ALWAYS and ASNEEDED are supported.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidQuoteFields,
-
-    /// The requested range cannot be satisfied.
-    ///
-    /// HTTP Status Code: 416 Requested Range NotSatisfiable
-    ///
-    InvalidRange,
-
-    /// You've attempted to create a Multi-Region Access Point in a Region that you haven't opted in to.
-    ///
-    /// HTTP Status Code: 403 Forbidden
-    ///
-    InvalidRegion,
-
-    /// + Please use <code>AWS4-HMAC-SHA256</code>.
-    /// + SOAP requests must be made over an HTTPS connection.
-    /// + Amazon S3 Transfer Acceleration is not supported for buckets with non-DNS compliant names.
-    /// + Amazon S3 Transfer Acceleration is not supported for buckets with periods (.) in their names.
-    /// + Amazon S3 Transfer Accelerate endpoint only supports virtual style requests.
-    /// + Amazon S3 Transfer Accelerate is not configured on this bucket.
-    /// + Amazon S3 Transfer Accelerate is disabled on this bucket.
-    /// + Amazon S3 Transfer Acceleration is not supported on this bucket. Contact Amazon Web Services Support for more information.
-    /// + Amazon S3 Transfer Acceleration cannot be enabled on this bucket. Contact Amazon Web Services Support for more information.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    InvalidRequest,
-
-    /// The value of a parameter in the SelectRequest element is not valid. Check the service API documentation and try again.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidRequestParameter,
-
-    /// The SOAP request body is invalid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidSOAPRequest,
-
-    /// The provided scan range is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidScanRange,
-
-    /// The provided security credentials are not valid.
-    ///
-    /// HTTP Status Code: 403 Forbidden
-    ///
-    InvalidSecurity,
-
-    /// Returned if the session doesn't exist anymore because it timed out or expired.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidSessionException,
-
-    /// The request signature that the server calculated does not match the signature that you provided. Check your AWS secret access key and signing method. For more information, see Signing and authenticating REST requests.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidSignature,
-
-    /// The storage class you specified is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidStorageClass,
-
-    /// The SQL expression contains a table alias that is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidTableAlias,
-
-    /// Your request contains tag input that is not valid. For example, your request might contain duplicate keys, keys or values that are too long, or system tags.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidTag,
-
-    /// The target bucket for logging does not exist, is not owned by you, or does not have the appropriate grants for the log-delivery group.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidTargetBucketForLogging,
-
-    /// The encoding type is not valid. Only UTF-8 encoding is supported.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidTextEncoding,
-
-    /// The provided token is malformed or otherwise invalid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidToken,
-
-    /// Couldn't parse the specified URI.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    InvalidURI,
-
-    /// An error occurred while parsing the JSON file. Check the file and try again.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    JSONParsingError,
-
-    /// Your key is too long.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    KeyTooLongError,
-
-    /// The SQL expression contains a character that is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    LexerInvalidChar,
-
-    /// The SQL expression contains an operator that is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    LexerInvalidIONLiteral,
-
-    /// The SQL expression contains an operator that is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    LexerInvalidLiteral,
-
-    /// The SQL expression contains a literal that is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    LexerInvalidOperator,
-
-    /// The argument given to the LIKE clause in the SQL expression is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    LikeInvalidInputs,
-
-    /// The XML you provided was not well-formed or did not validate against our published schema.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    MalformedACLError,
-
-    /// The body of your POST request is not well-formed multipart/form-data.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    MalformedPOSTRequest,
-
-    /// Your policy contains a principal that is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    MalformedPolicy,
-
-    /// This happens when the user sends malformed XML (XML that doesn't conform to the published XSD) for the configuration. The error message is, "The XML you provided was not well-formed or did not validate against our published schema."
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    MalformedXML,
-
-    /// Your request was too big.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    MaxMessageLengthExceeded,
-
-    /// Failed to parse SQL expression, try reducing complexity. For example, reduce number of operators used.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    MaxOperatorsExceeded,
-
-    /// Your POST request fields preceding the upload file were too large.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    MaxPostPreDataLengthExceededError,
-
-    /// Your metadata headers exceed the maximum allowed metadata size.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    MetadataTooLarge,
-
-    /// The specified method is not allowed against this resource.
-    ///
-    /// HTTP Status Code: 405 Method Not Allowed
-    ///
-    MethodNotAllowed,
-
-    /// A SOAP attachment was expected, but none were found.
-    ///
-    MissingAttachment,
-
-    /// The request was not signed.Â
-    ///
-    /// HTTP Status Code: 403 Forbidden
-    ///
-    MissingAuthenticationToken,
-
-    /// You must provide the Content-Length HTTP header.
-    ///
-    /// HTTP Status Code: 411 Length Required
-    ///
-    MissingContentLength,
-
-    /// This happens when the user sends an empty XML document as a request. The error message is, "Request body is empty."
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    MissingRequestBodyError,
-
-    /// The SelectRequest entity is missing a required parameter. Check the service documentation and try again.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    MissingRequiredParameter,
-
-    /// The SOAP 1.1 request is missing a security element.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    MissingSecurityElement,
-
-    /// Your request is missing a required header.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    MissingSecurityHeader,
-
-    /// Multiple data sources are not supported.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    MultipleDataSourcesUnsupported,
-
-    /// There is no such thing as a logging status subresource for a key.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    NoLoggingStatusForKey,
-
-    /// The specified access point does not exist.
-    ///
-    /// HTTP Status Code: 404 Not Found
-    ///
-    NoSuchAccessPoint,
-
-    /// The specified request was not found.
-    ///
-    /// HTTP Status Code: 404 Not Found
-    ///
-    NoSuchAsyncRequest,
-
-    /// The specified bucket does not exist.
-    ///
-    /// HTTP Status Code: 404 Not Found
-    ///
-    NoSuchBucket,
-
-    /// The specified bucket does not have a bucket policy.
-    ///
-    /// HTTP Status Code: 404 Not Found
-    ///
-    NoSuchBucketPolicy,
-
-    /// The specified bucket does not have a CORS configuration.
-    ///
-    /// HTTP Status Code: 404 Not Found
-    ///
-    NoSuchCORSConfiguration,
-
-    /// The specified key does not exist.
-    ///
-    /// HTTP Status Code: 404 Not Found
-    ///
-    NoSuchKey,
-
-    /// The lifecycle configuration does not exist.
-    ///
-    /// HTTP Status Code: 404 Not Found
-    ///
-    NoSuchLifecycleConfiguration,
-
-    /// The specified Multi-Region Access Point does not exist.
-    ///
-    /// HTTP Status Code: 404 Not Found
-    ///
-    NoSuchMultiRegionAccessPoint,
-
-    /// The specified object does not have an ObjectLock configuration.
-    ///
-    /// HTTP Status Code: 404 Not Found
-    ///
-    NoSuchObjectLockConfiguration,
-
-    /// The specified resource doesn't exist.
-    ///
-    /// HTTP Status Code: 404 Not Found
-    ///
-    NoSuchResource,
-
-    /// The specified tag does not exist.
-    ///
-    /// HTTP Status Code: 404 Not Found
-    ///
-    NoSuchTagSet,
-
-    /// The specified multipart upload does not exist. The upload ID might be invalid, or the multipart upload might have been aborted or completed.
-    ///
-    /// HTTP Status Code: 404 Not Found
-    ///
-    NoSuchUpload,
-
-    /// Indicates that the version ID specified in the request does not match an existing version.
-    ///
-    /// HTTP Status Code: 404 Not Found
-    ///
-    NoSuchVersion,
-
-    /// The specified bucket does not have a website configuration.
-    ///
-    /// HTTP Status Code: 404 Not Found
-    ///
-    NoSuchWebsiteConfiguration,
-
-    /// No transformation found for this Object Lambda Access Point.
-    ///
-    /// HTTP Status Code: 404 Not Found
-    ///
-    NoTransformationDefined,
-
-    /// The device that generated the token is not owned by the authenticated user.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    NotDeviceOwnerError,
-
-    /// A header you provided implies functionality that is not implemented.
-    ///
-    /// HTTP Status Code: 501 Not Implemented
-    ///
-    NotImplemented,
-
-    /// The resource was not changed.
-    ///
-    /// HTTP Status Code: 304 Not Modified
-    ///
-    NotModified,
-
-    /// Your account is not signed up for the Amazon S3 service. You must sign up before you can use Amazon S3. You can sign up at the following URL: <a href="http://aws.amazon.com/s3">Amazon S3</a>
-    ///
-    /// HTTP Status Code: 403 Forbidden
-    ///
-    NotSignedUp,
-
-    /// An error occurred while parsing a number. This error can be caused by underflow or overflow of integers.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    NumberFormatError,
-
-    /// The Object Lock configuration does not exist for this bucket.
-    ///
-    /// HTTP Status Code: 404 Not Found
-    ///
-    ObjectLockConfigurationNotFoundError,
-
-    /// InputSerialization specifies more than one format (CSV, JSON, or Parquet), or OutputSerialization specifies more than one format (CSV or JSON). For InputSerialization and OutputSerialization, you can specify only one format for each.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ObjectSerializationConflict,
-
-    /// A conflicting conditional action is currently in progress against this resource. Try again.
-    ///
-    /// HTTP Status Code: 409 Conflict
-    ///
-    OperationAborted,
-
-    /// The number of columns in the result is greater than the maximum allowable number of columns.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    OverMaxColumn,
-
-    /// The Parquet file is above the max row group size.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    OverMaxParquetBlockSize,
-
-    /// The length of a record in the input or result is greater than the maxCharsPerRecord limit of 1 MB.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    OverMaxRecordSize,
-
-    /// The bucket ownership controls were not found.
-    ///
-    /// HTTP Status Code: 404 Not Found
-    ///
-    OwnershipControlsNotFoundError,
-
-    /// An error occurred while parsing the Parquet file. Check the file and try again.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParquetParsingError,
-
-    /// The specified Parquet compression codec is not supported.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParquetUnsupportedCompressionCodec,
-
-    /// Other expressions are not allowed in the SELECT list when * is used without dot notation in the SQL expression.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseAsteriskIsNotAloneInSelectList,
-
-    /// Cannot mix [] and * in the same expression in a SELECT list in the SQL expression.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseCannotMixSqbAndWildcardInSelectList,
-
-    /// The SQL expression CAST has incorrect arity.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseCastArity,
-
-    /// The SQL expression contains an empty SELECT clause.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseEmptySelect,
-
-    /// The expected token in the SQL expression was not found.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseExpected2TokenTypes,
-
-    /// The expected argument delimiter in the SQL expression was not found.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseExpectedArgumentDelimiter,
-
-    /// The expected date part in the SQL expression was not found.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseExpectedDatePart,
-
-    /// The expected SQL expression was not found.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseExpectedExpression,
-
-    /// The expected identifier for the alias in the SQL expression was not found.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseExpectedIdentForAlias,
-
-    /// The expected identifier for AT name in the SQL expression was not found.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseExpectedIdentForAt,
-
-    /// GROUP is not supported in the SQL expression.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseExpectedIdentForGroupName,
-
-    /// The expected keyword in the SQL expression was not found.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseExpectedKeyword,
-
-    /// The expected left parenthesis after CAST in the SQL expression was not found.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseExpectedLeftParenAfterCast,
-
-    /// The expected left parenthesis in the SQL expression was not found.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseExpectedLeftParenBuiltinFunctionCall,
-
-    /// The expected left parenthesis in the SQL expression was not found.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseExpectedLeftParenValueConstructor,
-
-    /// The SQL expression contains an unsupported use of MEMBER.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseExpectedMember,
-
-    /// The expected number in the SQL expression was not found.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseExpectedNumber,
-
-    /// The expected right parenthesis character in the SQL expression was not found.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseExpectedRightParenBuiltinFunctionCall,
-
-    /// The expected token in the SQL expression was not found.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseExpectedTokenType,
-
-    /// The expected type name in the SQL expression was not found.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseExpectedTypeName,
-
-    /// The expected WHEN clause in the SQL expression was not found. CASE is not supported.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseExpectedWhenClause,
-
-    /// The use of * in the SELECT list in the SQL expression is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseInvalidContextForWildcardInSelectList,
-
-    /// The SQL expression contains a path component that is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseInvalidPathComponent,
-
-    /// The SQL expression contains a parameter value that is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseInvalidTypeParam,
-
-    /// JOIN is not supported in the SQL expression.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseMalformedJoin,
-
-    /// The expected identifier after the @ symbol in the SQL expression was not found.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseMissingIdentAfterAt,
-
-    /// Only one argument is supported for aggregate functions in the SQL expression.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseNonUnaryAgregateFunctionCall,
-
-    /// The SQL expression contains a missing FROM after the SELECT list.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseSelectMissingFrom,
-
-    /// The SQL expression contains an unexpected keyword.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseUnExpectedKeyword,
-
-    /// The SQL expression contains an unexpected operator.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseUnexpectedOperator,
-
-    /// The SQL expression contains an unexpected term.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseUnexpectedTerm,
-
-    /// The SQL expression contains an unexpected token.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseUnexpectedToken,
-
-    /// The SQL expression contains an operator that is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseUnknownOperator,
-
-    /// The SQL expression contains an unsupported use of ALIAS.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseUnsupportedAlias,
-
-    /// Only COUNT with (*) as a parameter is supported in the SQL expression.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseUnsupportedCallWithStar,
-
-    /// The SQL expression contains an unsupported use of CASE.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseUnsupportedCase,
-
-    /// The SQL expression contains an unsupported use of CASE.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseUnsupportedCaseClause,
-
-    /// The SQL expression contains an unsupported use of GROUP BY.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseUnsupportedLiteralsGroupBy,
-
-    /// The SQL expression contains an unsupported use of SELECT.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseUnsupportedSelect,
-
-    /// The SQL expression contains unsupported syntax.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseUnsupportedSyntax,
-
-    /// The SQL expression contains an unsupported token.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ParseUnsupportedToken,
-
-    /// The bucket you are attempting to access must be addressed using the specified endpoint. Send all future requests to this endpoint.
-    ///
-    /// HTTP Status Code: 301 Moved Permanently
-    ///
-    PermanentRedirect,
-
-    /// The API operation you are attempting to access must be addressed using the specified endpoint. Send all future requests to this endpoint.
-    ///
-    /// HTTP Status Code: 301 Moved Permanently
-    ///
-    PermanentRedirectControlError,
-
-    /// At least one of the preconditions you specified did not hold.
-    ///
-    /// HTTP Status Code: 412 Precondition Failed
-    ///
-    PreconditionFailed,
-
-    /// Temporary redirect.
-    ///
-    /// HTTP Status Code: 307 Moved Temporarily
-    ///
-    Redirect,
-
-    /// There is no replication configuration for this bucket.
-    ///
-    /// HTTP Status Code: 404 Not Found
-    ///
-    ReplicationConfigurationNotFoundError,
-
-    /// The request header and query parameters used to make the request exceed the maximum allowed size.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    RequestHeaderSectionTooLarge,
-
-    /// Bucket POST must be of the enclosure-type multipart/form-data.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    RequestIsNotMultiPartContent,
-
-    /// The difference between the request time and the server's time is too large.
-    ///
-    /// HTTP Status Code: 403 Forbidden
-    ///
-    RequestTimeTooSkewed,
-
-    /// Your socket connection to the server was not read from or written to within the timeout period.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    RequestTimeout,
-
-    /// Requesting the torrent file of a bucket is not permitted.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    RequestTorrentOfBucketError,
-
-    /// Returned to the original caller when an error is encountered while reading the WriteGetObjectResponse body.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ResponseInterrupted,
-
-    /// Object restore is already in progress.
-    ///
-    /// HTTP Status Code: 409 Conflict
-    ///
-    RestoreAlreadyInProgress,
-
-    /// The server-side encryption configuration was not found.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ServerSideEncryptionConfigurationNotFoundError,
-
-    /// Service is unable to handle request.
-    ///
-    /// HTTP Status Code: 503 Service Unavailable
-    ///
-    ServiceUnavailable,
-
-    /// The request signature we calculated does not match the signature you provided. Check your Amazon Web Services secret access key and signing method. For more information, see <a href="https://docs.aws.amazon.com/AmazonS3/latest/dev/RESTAuthentication.html">REST Authentication</a> and <a href="https://docs.aws.amazon.com/AmazonS3/latest/dev/SOAPAuthentication.html">SOAP Authentication</a> for details.
-    ///
-    /// HTTP Status Code: 403 Forbidden
-    ///
-    SignatureDoesNotMatch,
-
-    /// Reduce your request rate.
-    ///
-    /// HTTP Status Code: 503 Slow Down
-    ///
-    SlowDown,
-
-    /// The tag policy does not allow the specified value for the following tag key.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    TagPolicyException,
-
-    /// You are being redirected to the bucket while DNS updates.
-    ///
-    /// HTTP Status Code: 307 Moved Temporarily
-    ///
-    TemporaryRedirect,
-
-    /// The serial number and/or token code you provided is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    TokenCodeInvalidError,
-
-    /// The provided token must be refreshed.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    TokenRefreshRequired,
-
-    /// You have attempted to create more access points than are allowed for an account. For more information, see Amazon Simple Storage Service endpoints and quotas in the AWS General Reference.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    TooManyAccessPoints,
-
-    /// You have attempted to create more buckets than allowed.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    TooManyBuckets,
-
-    /// You have attempted to create a Multi-Region Access Point with more Regions than are allowed for an account. For more information, see Amazon Simple Storage Service endpoints and quotas in the AWS General Reference.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    TooManyMultiRegionAccessPointregionsError,
-
-    /// You have attempted to create more Multi-Region Access Points than are allowed for an account. For more information, see Amazon Simple Storage Service endpoints and quotas in the AWS General Reference.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    TooManyMultiRegionAccessPoints,
-
-    /// The number of tags exceeds the limit of 50 tags.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    TooManyTags,
-
-    /// Object decompression failed. Check that the object is properly compressed using the format specified in the request.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    TruncatedInput,
-
-    /// You are not authorized to perform this operation.
-    ///
-    /// HTTP Status Code: 401 Unauthorized
-    ///
-    UnauthorizedAccess,
-
-    /// Applicable in China Regions only. Returned when a request is made to a bucket that doesn't have an ICP license. For more information, see ICP Recordal.
-    ///
-    /// HTTP Status Code: 403 Forbidden
-    ///
-    UnauthorizedAccessError,
-
-    /// This request does not support content.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    UnexpectedContent,
-
-    /// Applicable in China Regions only. This request was rejected because the IP was unexpected.
-    ///
-    /// HTTP Status Code: 403 Forbidden
-    ///
-    UnexpectedIPError,
-
-    /// We encountered a record type that is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    UnrecognizedFormatException,
-
-    /// The email address you provided does not match any account on record.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    UnresolvableGrantByEmailAddress,
-
-    /// The request contained an unsupported argument.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    UnsupportedArgument,
-
-    /// We encountered an unsupported SQL function.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    UnsupportedFunction,
-
-    /// The specified Parquet type is not supported.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    UnsupportedParquetType,
-
-    /// A range header is not supported for this operation.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    UnsupportedRangeHeader,
-
-    /// Scan range queries are not supported on this type of object.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    UnsupportedScanRangeInput,
-
-    /// The provided request is signed with an unsupported STS Token version or the signature version is not supported.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    UnsupportedSignature,
-
-    /// We encountered an unsupported SQL operation.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    UnsupportedSqlOperation,
-
-    /// We encountered an unsupported SQL structure. Check the SQL Reference.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    UnsupportedSqlStructure,
-
-    /// We encountered a storage class that is not supported. Only STANDARD, STANDARD_IA, and ONEZONE_IA storage classes are supported.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    UnsupportedStorageClass,
-
-    /// We encountered syntax that is not valid.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    UnsupportedSyntax,
-
-    /// Your query contains an unsupported type for comparison (e.g. verifying that a Parquet INT96 column type is greater than 0).
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    UnsupportedTypeForQuerying,
-
-    /// The bucket POST must contain the specified field name. If it is specified, check the order of the fields.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    UserKeyMustBeSpecified,
-
-    /// A timestamp parse failure occurred in the SQL expression.
-    ///
-    /// HTTP Status Code: 400 Bad Request
-    ///
-    ValueParseFailure,
-
-    Custom(ByteString),
-}
-
-static S3_ERROR_CODE_MAP: phf::Map<&'static str, S3ErrorCode> = ::phf::Map {
-    key: 16287231350648472473,
-    disps: &[
-        (0, 4),
-        (0, 1),
-        (0, 10),
-        (0, 2),
-        (0, 2),
-        (0, 0),
-        (0, 0),
-        (0, 18),
-        (0, 53),
-        (0, 119),
-        (0, 27),
-        (0, 0),
-        (0, 12),
-        (0, 0),
-        (0, 1),
-        (0, 33),
-        (0, 100),
-        (1, 0),
-        (0, 37),
-        (0, 15),
-        (0, 1),
-        (0, 1),
-        (0, 1),
-        (0, 67),
-        (0, 11),
-        (0, 0),
-        (0, 10),
-        (1, 0),
-        (0, 14),
-        (0, 73),
-        (0, 9),
-        (0, 92),
-        (0, 2),
-        (0, 3),
-        (0, 2),
-        (0, 4),
-        (0, 48),
-        (0, 35),
-        (0, 11),
-        (0, 83),
-        (0, 48),
-        (0, 71),
-        (0, 10),
-        (0, 20),
-        (0, 82),
-        (0, 17),
-        (0, 29),
-        (0, 22),
-        (0, 4),
-        (0, 1),
-        (0, 111),
-        (0, 47),
-        (0, 99),
-        (0, 30),
-        (0, 39),
-        (0, 103),
-        (0, 32),
-        (1, 123),
-        (0, 209),
-        (0, 7),
-        (0, 9),
-        (0, 0),
-        (0, 54),
-        (0, 0),
-        (0, 43),
-        (0, 3),
-        (0, 1),
-        (0, 40),
-        (0, 18),
-        (0, 17),
-        (0, 79),
-        (0, 3),
-        (0, 46),
-        (0, 114),
-        (0, 2),
-        (0, 103),
-        (0, 2),
-        (0, 114),
-        (0, 112),
-        (0, 3),
-    ],
-    entries: &[
-        ("ParseUnExpectedKeyword", S3ErrorCode::ParseUnExpectedKeyword),
-        ("MaxMessageLengthExceeded", S3ErrorCode::MaxMessageLengthExceeded),
-        ("CSVParsingError", S3ErrorCode::CSVParsingError),
-        ("EmptyRequestBody", S3ErrorCode::EmptyRequestBody),
-        ("AccessControlListNotSupported", S3ErrorCode::AccessControlListNotSupported),
-        ("MalformedPolicy", S3ErrorCode::MalformedPolicy),
-        ("ParseAsteriskIsNotAloneInSelectList", S3ErrorCode::ParseAsteriskIsNotAloneInSelectList),
-        ("InvalidAddressingHeader", S3ErrorCode::InvalidAddressingHeader),
-        ("RequestHeaderSectionTooLarge", S3ErrorCode::RequestHeaderSectionTooLarge),
-        ("MalformedPOSTRequest", S3ErrorCode::MalformedPOSTRequest),
-        ("OperationAborted", S3ErrorCode::OperationAborted),
-        ("AuthorizationHeaderMalformed", S3ErrorCode::AuthorizationHeaderMalformed),
-        ("InvalidEncryptionAlgorithmError", S3ErrorCode::InvalidEncryptionAlgorithmError),
-        ("IllegalLocationConstraintException", S3ErrorCode::IllegalLocationConstraintException),
-        ("InvalidStorageClass", S3ErrorCode::InvalidStorageClass),
-        ("InvalidObjectState", S3ErrorCode::InvalidObjectState),
-        ("UnsupportedFunction", S3ErrorCode::UnsupportedFunction),
-        ("UnsupportedRangeHeader", S3ErrorCode::UnsupportedRangeHeader),
-        ("ParseExpectedTypeName", S3ErrorCode::ParseExpectedTypeName),
-        ("PermanentRedirectControlError", S3ErrorCode::PermanentRedirectControlError),
-        ("CSVEscapingRecordDelimiter", S3ErrorCode::CSVEscapingRecordDelimiter),
-        ("AmbiguousFieldName", S3ErrorCode::AmbiguousFieldName),
-        ("InvalidBucketName", S3ErrorCode::InvalidBucketName),
-        (
-            "EvaluatorInvalidTimestampFormatPattern",
-            S3ErrorCode::EvaluatorInvalidTimestampFormatPattern,
-        ),
-        ("EntityTooSmall", S3ErrorCode::EntityTooSmall),
-        ("InvalidURI", S3ErrorCode::InvalidURI),
-        ("MissingSecurityHeader", S3ErrorCode::MissingSecurityHeader),
-        ("InvalidTableAlias", S3ErrorCode::InvalidTableAlias),
-        ("UnsupportedScanRangeInput", S3ErrorCode::UnsupportedScanRangeInput),
-        ("PermanentRedirect", S3ErrorCode::PermanentRedirect),
-        ("AllAccessDisabled", S3ErrorCode::AllAccessDisabled),
-        ("UnsupportedSqlOperation", S3ErrorCode::UnsupportedSqlOperation),
-        ("NotImplemented", S3ErrorCode::NotImplemented),
-        ("InvalidPayer", S3ErrorCode::InvalidPayer),
-        ("InvalidDataType", S3ErrorCode::InvalidDataType),
-        ("InvalidPolicyDocument", S3ErrorCode::InvalidPolicyDocument),
-        ("IncorrectEndpoint", S3ErrorCode::IncorrectEndpoint),
-        ("ParseExpectedArgumentDelimiter", S3ErrorCode::ParseExpectedArgumentDelimiter),
-        ("LexerInvalidChar", S3ErrorCode::LexerInvalidChar),
-        (
-            "EvaluatorInvalidTimestampFormatPatternSymbolForParsing",
-            S3ErrorCode::EvaluatorInvalidTimestampFormatPatternSymbolForParsing,
-        ),
-        ("ConnectionClosedByRequester", S3ErrorCode::ConnectionClosedByRequester),
-        ("InvalidDataSource", S3ErrorCode::InvalidDataSource),
-        (
-            "EvaluatorInvalidTimestampFormatPatternSymbol",
-            S3ErrorCode::EvaluatorInvalidTimestampFormatPatternSymbol,
-        ),
-        ("InvalidJsonType", S3ErrorCode::InvalidJsonType),
-        ("NoSuchLifecycleConfiguration", S3ErrorCode::NoSuchLifecycleConfiguration),
-        ("IncompleteBody", S3ErrorCode::IncompleteBody),
-        ("ParseUnexpectedToken", S3ErrorCode::ParseUnexpectedToken),
-        ("UnsupportedTypeForQuerying", S3ErrorCode::UnsupportedTypeForQuerying),
-        ("InvalidQuoteFields", S3ErrorCode::InvalidQuoteFields),
-        ("UserKeyMustBeSpecified", S3ErrorCode::UserKeyMustBeSpecified),
-        ("ClientTokenConflict", S3ErrorCode::ClientTokenConflict),
-        ("EndpointNotFound", S3ErrorCode::EndpointNotFound),
-        ("OwnershipControlsNotFoundError", S3ErrorCode::OwnershipControlsNotFoundError),
-        ("ExpiredToken", S3ErrorCode::ExpiredToken),
-        ("ParseUnsupportedSyntax", S3ErrorCode::ParseUnsupportedSyntax),
-        ("InvalidArgument", S3ErrorCode::InvalidArgument),
-        ("MaxOperatorsExceeded", S3ErrorCode::MaxOperatorsExceeded),
-        ("CastFailed", S3ErrorCode::CastFailed),
-        ("ExternalEvalException", S3ErrorCode::ExternalEvalException),
-        ("NoSuchBucketPolicy", S3ErrorCode::NoSuchBucketPolicy),
-        ("InvalidHostHeader", S3ErrorCode::InvalidHostHeader),
-        ("LexerInvalidLiteral", S3ErrorCode::LexerInvalidLiteral),
-        ("InvalidSignature", S3ErrorCode::InvalidSignature),
-        ("ParseUnsupportedSelect", S3ErrorCode::ParseUnsupportedSelect),
-        ("ParseExpectedDatePart", S3ErrorCode::ParseExpectedDatePart),
-        ("InvalidBucketState", S3ErrorCode::InvalidBucketState),
-        ("TruncatedInput", S3ErrorCode::TruncatedInput),
-        ("MissingContentLength", S3ErrorCode::MissingContentLength),
-        ("ParseInvalidPathComponent", S3ErrorCode::ParseInvalidPathComponent),
-        ("ParseExpectedIdentForGroupName", S3ErrorCode::ParseExpectedIdentForGroupName),
-        ("DeviceNotActiveError", S3ErrorCode::DeviceNotActiveError),
-        ("Redirect", S3ErrorCode::Redirect),
-        ("LexerInvalidOperator", S3ErrorCode::LexerInvalidOperator),
-        ("InvalidTargetBucketForLogging", S3ErrorCode::InvalidTargetBucketForLogging),
-        ("MethodNotAllowed", S3ErrorCode::MethodNotAllowed),
-        ("MissingAttachment", S3ErrorCode::MissingAttachment),
-        ("InvalidRequest", S3ErrorCode::InvalidRequest),
-        ("InvalidSecurity", S3ErrorCode::InvalidSecurity),
-        ("ExpressionTooLong", S3ErrorCode::ExpressionTooLong),
-        ("ParseNonUnaryAgregateFunctionCall", S3ErrorCode::ParseNonUnaryAgregateFunctionCall),
-        ("ParseInvalidTypeParam", S3ErrorCode::ParseInvalidTypeParam),
-        ("MalformedXML", S3ErrorCode::MalformedXML),
-        ("ParseExpectedKeyword", S3ErrorCode::ParseExpectedKeyword),
-        ("RequestTorrentOfBucketError", S3ErrorCode::RequestTorrentOfBucketError),
-        ("UnexpectedIPError", S3ErrorCode::UnexpectedIPError),
-        ("InvalidCast", S3ErrorCode::InvalidCast),
-        (
-            "ParseExpectedRightParenBuiltinFunctionCall",
-            S3ErrorCode::ParseExpectedRightParenBuiltinFunctionCall,
-        ),
-        ("ParseExpectedTokenType", S3ErrorCode::ParseExpectedTokenType),
-        ("AccountProblem", S3ErrorCode::AccountProblem),
-        ("AccessDenied", S3ErrorCode::AccessDenied),
-        ("InvalidAccessPointAliasError", S3ErrorCode::InvalidAccessPointAliasError),
-        ("InvalidColumnIndex", S3ErrorCode::InvalidColumnIndex),
-        ("InvalidDigest", S3ErrorCode::InvalidDigest),
-        ("NotSignedUp", S3ErrorCode::NotSignedUp),
-        ("ParseExpectedIdentForAlias", S3ErrorCode::ParseExpectedIdentForAlias),
-        ("ObjectSerializationConflict", S3ErrorCode::ObjectSerializationConflict),
-        ("ParseUnexpectedOperator", S3ErrorCode::ParseUnexpectedOperator),
-        ("TokenCodeInvalidError", S3ErrorCode::TokenCodeInvalidError),
-        ("ParseExpectedExpression", S3ErrorCode::ParseExpectedExpression),
-        ("ParseExpectedMember", S3ErrorCode::ParseExpectedMember),
-        ("PreconditionFailed", S3ErrorCode::PreconditionFailed),
-        ("BadDigest", S3ErrorCode::BadDigest),
-        ("InvalidPart", S3ErrorCode::InvalidPart),
-        ("InvalidLocationConstraint", S3ErrorCode::InvalidLocationConstraint),
-        ("ParquetUnsupportedCompressionCodec", S3ErrorCode::ParquetUnsupportedCompressionCodec),
-        ("ParseCastArity", S3ErrorCode::ParseCastArity),
-        ("NoTransformationDefined", S3ErrorCode::NoTransformationDefined),
-        ("InvalidCompressionFormat", S3ErrorCode::InvalidCompressionFormat),
-        ("NoSuchAccessPoint", S3ErrorCode::NoSuchAccessPoint),
-        ("InlineDataTooLarge", S3ErrorCode::InlineDataTooLarge),
-        ("TooManyMultiRegionAccessPoints", S3ErrorCode::TooManyMultiRegionAccessPoints),
-        ("ParseUnsupportedCallWithStar", S3ErrorCode::ParseUnsupportedCallWithStar),
-        ("ParseUnexpectedTerm", S3ErrorCode::ParseUnexpectedTerm),
-        ("Busy", S3ErrorCode::Busy),
-        ("InvalidTextEncoding", S3ErrorCode::InvalidTextEncoding),
-        ("IncorrectSqlFunctionArgumentType", S3ErrorCode::IncorrectSqlFunctionArgumentType),
-        ("UnresolvableGrantByEmailAddress", S3ErrorCode::UnresolvableGrantByEmailAddress),
-        ("NoSuchVersion", S3ErrorCode::NoSuchVersion),
-        ("InvalidRange", S3ErrorCode::InvalidRange),
-        ("InvalidPartOrder", S3ErrorCode::InvalidPartOrder),
-        ("BucketNotEmpty", S3ErrorCode::BucketNotEmpty),
-        ("TemporaryRedirect", S3ErrorCode::TemporaryRedirect),
-        ("MissingSecurityElement", S3ErrorCode::MissingSecurityElement),
-        ("TooManyAccessPoints", S3ErrorCode::TooManyAccessPoints),
-        ("UnauthorizedAccessError", S3ErrorCode::UnauthorizedAccessError),
-        ("NoSuchWebsiteConfiguration", S3ErrorCode::NoSuchWebsiteConfiguration),
-        ("ParseExpectedWhenClause", S3ErrorCode::ParseExpectedWhenClause),
-        ("RestoreAlreadyInProgress", S3ErrorCode::RestoreAlreadyInProgress),
-        ("MissingAuthenticationToken", S3ErrorCode::MissingAuthenticationToken),
-        ("ParseExpected2TokenTypes", S3ErrorCode::ParseExpected2TokenTypes),
-        ("AccessPointAlreadyOwnedByYou", S3ErrorCode::AccessPointAlreadyOwnedByYou),
-        ("UnsupportedSignature", S3ErrorCode::UnsupportedSignature),
-        ("ParseUnsupportedCaseClause", S3ErrorCode::ParseUnsupportedCaseClause),
-        (
-            "TooManyMultiRegionAccessPointregionsError",
-            S3ErrorCode::TooManyMultiRegionAccessPointregionsError,
-        ),
-        ("NumberFormatError", S3ErrorCode::NumberFormatError),
-        (
-            "ParseExpectedLeftParenValueConstructor",
-            S3ErrorCode::ParseExpectedLeftParenValueConstructor,
-        ),
-        ("MissingRequiredParameter", S3ErrorCode::MissingRequiredParameter),
-        ("NoSuchObjectLockConfiguration", S3ErrorCode::NoSuchObjectLockConfiguration),
-        ("NoSuchCORSConfiguration", S3ErrorCode::NoSuchCORSConfiguration),
-        (
-            "ReplicationConfigurationNotFoundError",
-            S3ErrorCode::ReplicationConfigurationNotFoundError,
-        ),
-        ("ParseEmptySelect", S3ErrorCode::ParseEmptySelect),
-        ("AuthorizationQueryParametersError", S3ErrorCode::AuthorizationQueryParametersError),
-        ("SignatureDoesNotMatch", S3ErrorCode::SignatureDoesNotMatch),
-        ("TooManyTags", S3ErrorCode::TooManyTags),
-        ("ColumnTooLong", S3ErrorCode::ColumnTooLong),
-        ("JSONParsingError", S3ErrorCode::JSONParsingError),
-        ("BucketAlreadyOwnedByYou", S3ErrorCode::BucketAlreadyOwnedByYou),
-        ("BucketAlreadyExists", S3ErrorCode::BucketAlreadyExists),
-        ("AmbiguousGrantByEmailAddress", S3ErrorCode::AmbiguousGrantByEmailAddress),
-        (
-            "ServerSideEncryptionConfigurationNotFoundError",
-            S3ErrorCode::ServerSideEncryptionConfigurationNotFoundError,
-        ),
-        ("InvalidFileHeaderInfo", S3ErrorCode::InvalidFileHeaderInfo),
-        ("TooManyBuckets", S3ErrorCode::TooManyBuckets),
-        ("ObjectLockConfigurationNotFoundError", S3ErrorCode::ObjectLockConfigurationNotFoundError),
-        ("ResponseInterrupted", S3ErrorCode::ResponseInterrupted),
-        ("NoSuchMultiRegionAccessPoint", S3ErrorCode::NoSuchMultiRegionAccessPoint),
-        ("BucketHasAccessPointsAttached", S3ErrorCode::BucketHasAccessPointsAttached),
-        (
-            "EvaluatorTimestampFormatPatternDuplicateFields",
-            S3ErrorCode::EvaluatorTimestampFormatPatternDuplicateFields,
-        ),
-        ("InvalidBucketOwnerAWSAccountID", S3ErrorCode::InvalidBucketOwnerAWSAccountID),
-        (
-            "IllegalVersioningConfigurationException",
-            S3ErrorCode::IllegalVersioningConfigurationException,
-        ),
-        ("EvaluatorBindingDoesNotExist", S3ErrorCode::EvaluatorBindingDoesNotExist),
-        ("MalformedACLError", S3ErrorCode::MalformedACLError),
-        (
-            "ParseCannotMixSqbAndWildcardInSelectList",
-            S3ErrorCode::ParseCannotMixSqbAndWildcardInSelectList,
-        ),
-        ("UnsupportedParquetType", S3ErrorCode::UnsupportedParquetType),
-        ("ParseUnsupportedToken", S3ErrorCode::ParseUnsupportedToken),
-        ("InvalidTag", S3ErrorCode::InvalidTag),
-        ("UnsupportedSqlStructure", S3ErrorCode::UnsupportedSqlStructure),
-        ("NotDeviceOwnerError", S3ErrorCode::NotDeviceOwnerError),
-        ("CredentialsNotSupported", S3ErrorCode::CredentialsNotSupported),
-        ("EvaluatorInvalidArguments", S3ErrorCode::EvaluatorInvalidArguments),
-        ("NoSuchBucket", S3ErrorCode::NoSuchBucket),
-        ("TagPolicyException", S3ErrorCode::TagPolicyException),
-        (
-            "EvaluatorTimestampFormatPatternHourClockAmPmMismatch",
-            S3ErrorCode::EvaluatorTimestampFormatPatternHourClockAmPmMismatch,
-        ),
-        ("InvalidScanRange", S3ErrorCode::InvalidScanRange),
-        ("ParseMissingIdentAfterAt", S3ErrorCode::ParseMissingIdentAfterAt),
-        ("MetadataTooLarge", S3ErrorCode::MetadataTooLarge),
-        (
-            "EvaluatorInvalidTimestampFormatPatternToken",
-            S3ErrorCode::EvaluatorInvalidTimestampFormatPatternToken,
-        ),
-        ("ParseMalformedJoin", S3ErrorCode::ParseMalformedJoin),
-        ("ParseExpectedNumber", S3ErrorCode::ParseExpectedNumber),
-        ("SlowDown", S3ErrorCode::SlowDown),
-        ("InvalidExpressionType", S3ErrorCode::InvalidExpressionType),
-        (
-            "ParseExpectedLeftParenBuiltinFunctionCall",
-            S3ErrorCode::ParseExpectedLeftParenBuiltinFunctionCall,
-        ),
-        ("ValueParseFailure", S3ErrorCode::ValueParseFailure),
-        ("ParseUnsupportedCase", S3ErrorCode::ParseUnsupportedCase),
-        ("ParseExpectedIdentForAt", S3ErrorCode::ParseExpectedIdentForAt),
-        ("InvalidAccessKeyId", S3ErrorCode::InvalidAccessKeyId),
-        (
-            "EvaluatorUnterminatedTimestampFormatPatternToken",
-            S3ErrorCode::EvaluatorUnterminatedTimestampFormatPatternToken,
-        ),
-        ("NoSuchAsyncRequest", S3ErrorCode::NoSuchAsyncRequest),
-        ("NotModified", S3ErrorCode::NotModified),
-        ("ServiceUnavailable", S3ErrorCode::ServiceUnavailable),
-        ("IllegalSqlFunctionArgument", S3ErrorCode::IllegalSqlFunctionArgument),
-        ("UnsupportedArgument", S3ErrorCode::UnsupportedArgument),
-        ("UnauthorizedAccess", S3ErrorCode::UnauthorizedAccess),
-        ("CSVUnescapedQuote", S3ErrorCode::CSVUnescapedQuote),
-        ("TokenRefreshRequired", S3ErrorCode::TokenRefreshRequired),
-        ("CrossLocationLoggingProhibited", S3ErrorCode::CrossLocationLoggingProhibited),
-        (
-            "ParseInvalidContextForWildcardInSelectList",
-            S3ErrorCode::ParseInvalidContextForWildcardInSelectList,
-        ),
-        ("InvalidAccessPoint", S3ErrorCode::InvalidAccessPoint),
-        ("UnsupportedStorageClass", S3ErrorCode::UnsupportedStorageClass),
-        ("NoSuchUpload", S3ErrorCode::NoSuchUpload),
-        ("RequestIsNotMultiPartContent", S3ErrorCode::RequestIsNotMultiPartContent),
-        ("InvalidSessionException", S3ErrorCode::InvalidSessionException),
-        (
-            "EvaluatorLikePatternInvalidEscapeSequence",
-            S3ErrorCode::EvaluatorLikePatternInvalidEscapeSequence,
-        ),
-        ("UnexpectedContent", S3ErrorCode::UnexpectedContent),
-        ("NoLoggingStatusForKey", S3ErrorCode::NoLoggingStatusForKey),
-        ("ParseUnknownOperator", S3ErrorCode::ParseUnknownOperator),
-        ("EvaluatorNegativeLimit", S3ErrorCode::EvaluatorNegativeLimit),
-        ("OverMaxColumn", S3ErrorCode::OverMaxColumn),
-        ("ConditionalRequestConflict", S3ErrorCode::ConditionalRequestConflict),
-        ("InvalidBucketAclWithObjectOwnership", S3ErrorCode::InvalidBucketAclWithObjectOwnership),
-        ("InvalidRegion", S3ErrorCode::InvalidRegion),
-        ("InvalidSOAPRequest", S3ErrorCode::InvalidSOAPRequest),
-        ("InvalidKeyPath", S3ErrorCode::InvalidKeyPath),
-        ("InternalError", S3ErrorCode::InternalError),
-        ("OverMaxParquetBlockSize", S3ErrorCode::OverMaxParquetBlockSize),
-        ("UnrecognizedFormatException", S3ErrorCode::UnrecognizedFormatException),
-        ("LexerInvalidIONLiteral", S3ErrorCode::LexerInvalidIONLiteral),
-        ("InvalidHttpMethod", S3ErrorCode::InvalidHttpMethod),
-        ("ParquetParsingError", S3ErrorCode::ParquetParsingError),
-        ("IncorrectNumberOfFilesInPostRequest", S3ErrorCode::IncorrectNumberOfFilesInPostRequest),
-        ("ParseExpectedLeftParenAfterCast", S3ErrorCode::ParseExpectedLeftParenAfterCast),
-        ("KeyTooLongError", S3ErrorCode::KeyTooLongError),
-        ("OverMaxRecordSize", S3ErrorCode::OverMaxRecordSize),
-        ("RequestTimeout", S3ErrorCode::RequestTimeout),
-        ("LikeInvalidInputs", S3ErrorCode::LikeInvalidInputs),
-        ("InvalidToken", S3ErrorCode::InvalidToken),
-        ("ParseUnsupportedLiteralsGroupBy", S3ErrorCode::ParseUnsupportedLiteralsGroupBy),
-        ("IntegerOverflow", S3ErrorCode::IntegerOverflow),
-        ("UnsupportedSyntax", S3ErrorCode::UnsupportedSyntax),
-        ("MaxPostPreDataLengthExceededError", S3ErrorCode::MaxPostPreDataLengthExceededError),
-        ("RequestTimeTooSkewed", S3ErrorCode::RequestTimeTooSkewed),
-        ("InvalidRequestParameter", S3ErrorCode::InvalidRequestParameter),
-        ("NoSuchTagSet", S3ErrorCode::NoSuchTagSet),
-        ("MissingRequestBodyError", S3ErrorCode::MissingRequestBodyError),
-        ("EntityTooLarge", S3ErrorCode::EntityTooLarge),
-        ("MultipleDataSourcesUnsupported", S3ErrorCode::MultipleDataSourcesUnsupported),
-        ("NoSuchResource", S3ErrorCode::NoSuchResource),
-        ("NoSuchKey", S3ErrorCode::NoSuchKey),
-        ("ParseUnsupportedAlias", S3ErrorCode::ParseUnsupportedAlias),
-        ("ParseSelectMissingFrom", S3ErrorCode::ParseSelectMissingFrom),
-    ],
-};
-
-static S3_ERROR_CODE_LOWERCASE_MAP: phf::Map<&'static str, S3ErrorCode> = ::phf::Map {
-    key: 16287231350648472473,
-    disps: &[
-        (0, 19),
-        (0, 28),
-        (0, 3),
-        (0, 38),
-        (0, 49),
-        (0, 1),
-        (0, 1),
-        (0, 6),
-        (0, 53),
-        (0, 0),
-        (0, 0),
-        (0, 0),
-        (0, 11),
-        (0, 1),
-        (0, 10),
-        (0, 4),
-        (0, 2),
-        (0, 10),
-        (0, 10),
-        (0, 1),
-        (0, 1),
-        (0, 18),
-        (0, 2),
-        (0, 0),
-        (0, 45),
-        (0, 11),
-        (0, 55),
-        (0, 9),
-        (0, 1),
-        (0, 136),
-        (0, 0),
-        (0, 17),
-        (0, 0),
-        (0, 0),
-        (0, 0),
-        (0, 58),
-        (0, 1),
-        (0, 9),
-        (0, 8),
-        (0, 67),
-        (0, 99),
-        (0, 4),
-        (0, 127),
-        (0, 9),
-        (0, 10),
-        (0, 31),
-        (0, 4),
-        (0, 21),
-        (0, 9),
-        (0, 1),
-        (0, 0),
-        (0, 0),
-        (0, 70),
-        (0, 60),
-        (0, 1),
-        (0, 69),
-        (0, 50),
-        (0, 5),
-        (0, 0),
-        (0, 1),
-        (0, 12),
-        (0, 0),
-        (3, 143),
-        (0, 11),
-        (0, 180),
-        (0, 161),
-        (0, 5),
-        (0, 36),
-        (0, 10),
-        (1, 0),
-        (1, 188),
-        (0, 10),
-        (0, 10),
-        (0, 211),
-        (0, 32),
-        (0, 19),
-        (1, 34),
-        (0, 0),
-        (2, 187),
-        (1, 14),
-    ],
-    entries: &[
-        ("bucketalreadyownedbyyou", S3ErrorCode::BucketAlreadyOwnedByYou),
-        ("signaturedoesnotmatch", S3ErrorCode::SignatureDoesNotMatch),
-        ("parseunexpectedterm", S3ErrorCode::ParseUnexpectedTerm),
-        ("conditionalrequestconflict", S3ErrorCode::ConditionalRequestConflict),
-        ("evaluatornegativelimit", S3ErrorCode::EvaluatorNegativeLimit),
-        (
-            "parsecannotmixsqbandwildcardinselectlist",
-            S3ErrorCode::ParseCannotMixSqbAndWildcardInSelectList,
-        ),
-        ("invalidtargetbucketforlogging", S3ErrorCode::InvalidTargetBucketForLogging),
-        ("nosuchtagset", S3ErrorCode::NoSuchTagSet),
-        ("userkeymustbespecified", S3ErrorCode::UserKeyMustBeSpecified),
-        ("accesscontrollistnotsupported", S3ErrorCode::AccessControlListNotSupported),
-        ("lexerinvalidionliteral", S3ErrorCode::LexerInvalidIONLiteral),
-        ("missingattachment", S3ErrorCode::MissingAttachment),
-        ("incorrectsqlfunctionargumenttype", S3ErrorCode::IncorrectSqlFunctionArgumentType),
-        ("multipledatasourcesunsupported", S3ErrorCode::MultipleDataSourcesUnsupported),
-        ("restorealreadyinprogress", S3ErrorCode::RestoreAlreadyInProgress),
-        ("invalidscanrange", S3ErrorCode::InvalidScanRange),
-        ("invaliddatatype", S3ErrorCode::InvalidDataType),
-        ("expiredtoken", S3ErrorCode::ExpiredToken),
-        ("parseunsupportedalias", S3ErrorCode::ParseUnsupportedAlias),
-        ("parseunexpectedkeyword", S3ErrorCode::ParseUnExpectedKeyword),
-        ("permanentredirectcontrolerror", S3ErrorCode::PermanentRedirectControlError),
-        ("ownershipcontrolsnotfounderror", S3ErrorCode::OwnershipControlsNotFoundError),
-        ("nosuchresource", S3ErrorCode::NoSuchResource),
-        ("internalerror", S3ErrorCode::InternalError),
-        ("parquetunsupportedcompressioncodec", S3ErrorCode::ParquetUnsupportedCompressionCodec),
-        ("valueparsefailure", S3ErrorCode::ValueParseFailure),
-        ("nosuchwebsiteconfiguration", S3ErrorCode::NoSuchWebsiteConfiguration),
-        ("invalidhttpmethod", S3ErrorCode::InvalidHttpMethod),
-        ("unsupportedargument", S3ErrorCode::UnsupportedArgument),
-        ("parsecastarity", S3ErrorCode::ParseCastArity),
-        ("castfailed", S3ErrorCode::CastFailed),
-        ("toomanybuckets", S3ErrorCode::TooManyBuckets),
-        ("invaliduri", S3ErrorCode::InvalidURI),
-        ("parseexpectedtypename", S3ErrorCode::ParseExpectedTypeName),
-        ("overmaxcolumn", S3ErrorCode::OverMaxColumn),
-        ("lexerinvalidchar", S3ErrorCode::LexerInvalidChar),
-        ("parseunsupportedsyntax", S3ErrorCode::ParseUnsupportedSyntax),
-        ("invalidpart", S3ErrorCode::InvalidPart),
-        ("invalidaddressingheader", S3ErrorCode::InvalidAddressingHeader),
-        ("unresolvablegrantbyemailaddress", S3ErrorCode::UnresolvableGrantByEmailAddress),
-        ("parseunsupportedcallwithstar", S3ErrorCode::ParseUnsupportedCallWithStar),
-        ("incorrectendpoint", S3ErrorCode::IncorrectEndpoint),
-        ("invalidsessionexception", S3ErrorCode::InvalidSessionException),
-        ("nosuchobjectlockconfiguration", S3ErrorCode::NoSuchObjectLockConfiguration),
-        ("columntoolong", S3ErrorCode::ColumnTooLong),
-        ("buckethasaccesspointsattached", S3ErrorCode::BucketHasAccessPointsAttached),
-        ("unexpectedcontent", S3ErrorCode::UnexpectedContent),
-        ("invalidtag", S3ErrorCode::InvalidTag),
-        ("numberformaterror", S3ErrorCode::NumberFormatError),
-        ("parseexpectedmember", S3ErrorCode::ParseExpectedMember),
-        ("parseunsupportedliteralsgroupby", S3ErrorCode::ParseUnsupportedLiteralsGroupBy),
-        ("invalidhostheader", S3ErrorCode::InvalidHostHeader),
-        ("parseunsupportedtoken", S3ErrorCode::ParseUnsupportedToken),
-        ("parseunsupportedselect", S3ErrorCode::ParseUnsupportedSelect),
-        ("invalidbucketaclwithobjectownership", S3ErrorCode::InvalidBucketAclWithObjectOwnership),
-        ("malformedpolicy", S3ErrorCode::MalformedPolicy),
-        ("nosuchaccesspoint", S3ErrorCode::NoSuchAccessPoint),
-        ("likeinvalidinputs", S3ErrorCode::LikeInvalidInputs),
-        ("invalidtoken", S3ErrorCode::InvalidToken),
-        (
-            "evaluatorlikepatterninvalidescapesequence",
-            S3ErrorCode::EvaluatorLikePatternInvalidEscapeSequence,
-        ),
-        (
-            "toomanymultiregionaccesspointregionserror",
-            S3ErrorCode::TooManyMultiRegionAccessPointregionsError,
-        ),
-        ("baddigest", S3ErrorCode::BadDigest),
-        ("unauthorizedaccesserror", S3ErrorCode::UnauthorizedAccessError),
-        ("unsupportedsignature", S3ErrorCode::UnsupportedSignature),
-        ("lexerinvalidoperator", S3ErrorCode::LexerInvalidOperator),
-        ("invalidjsontype", S3ErrorCode::InvalidJsonType),
-        ("missingrequestbodyerror", S3ErrorCode::MissingRequestBodyError),
-        ("temporaryredirect", S3ErrorCode::TemporaryRedirect),
-        ("unsupportedsqloperation", S3ErrorCode::UnsupportedSqlOperation),
-        ("unauthorizedaccess", S3ErrorCode::UnauthorizedAccess),
-        ("entitytoolarge", S3ErrorCode::EntityTooLarge),
-        ("metadatatoolarge", S3ErrorCode::MetadataTooLarge),
-        ("inlinedatatoolarge", S3ErrorCode::InlineDataTooLarge),
-        ("unsupportedfunction", S3ErrorCode::UnsupportedFunction),
-        ("parseexpectedtokentype", S3ErrorCode::ParseExpectedTokenType),
-        ("invalidcolumnindex", S3ErrorCode::InvalidColumnIndex),
-        ("invalidrequest", S3ErrorCode::InvalidRequest),
-        ("ambiguousgrantbyemailaddress", S3ErrorCode::AmbiguousGrantByEmailAddress),
-        ("parseemptyselect", S3ErrorCode::ParseEmptySelect),
-        ("parseunknownoperator", S3ErrorCode::ParseUnknownOperator),
-        ("invaliddatasource", S3ErrorCode::InvalidDataSource),
-        ("nosuchbucketpolicy", S3ErrorCode::NoSuchBucketPolicy),
-        ("parquetparsingerror", S3ErrorCode::ParquetParsingError),
-        ("clienttokenconflict", S3ErrorCode::ClientTokenConflict),
-        ("invalidrequestparameter", S3ErrorCode::InvalidRequestParameter),
-        ("requesttimeout", S3ErrorCode::RequestTimeout),
-        ("unsupportedparquettype", S3ErrorCode::UnsupportedParquetType),
-        ("invalidregion", S3ErrorCode::InvalidRegion),
-        ("authorizationheadermalformed", S3ErrorCode::AuthorizationHeaderMalformed),
-        ("invalidkeypath", S3ErrorCode::InvalidKeyPath),
-        ("notsignedup", S3ErrorCode::NotSignedUp),
-        ("evaluatorinvalidarguments", S3ErrorCode::EvaluatorInvalidArguments),
-        ("unsupportedsyntax", S3ErrorCode::UnsupportedSyntax),
-        ("invalidobjectstate", S3ErrorCode::InvalidObjectState),
-        ("notdeviceownererror", S3ErrorCode::NotDeviceOwnerError),
-        ("parsemissingidentafterat", S3ErrorCode::ParseMissingIdentAfterAt),
-        (
-            "evaluatortimestampformatpatternhourclockampmmismatch",
-            S3ErrorCode::EvaluatorTimestampFormatPatternHourClockAmPmMismatch,
-        ),
-        ("invalidbucketstate", S3ErrorCode::InvalidBucketState),
-        ("accesspointalreadyownedbyyou", S3ErrorCode::AccessPointAlreadyOwnedByYou),
-        ("keytoolongerror", S3ErrorCode::KeyTooLongError),
-        ("unsupportedtypeforquerying", S3ErrorCode::UnsupportedTypeForQuerying),
-        ("invalidexpressiontype", S3ErrorCode::InvalidExpressionType),
-        ("parsemalformedjoin", S3ErrorCode::ParseMalformedJoin),
-        ("invalidaccesskeyid", S3ErrorCode::InvalidAccessKeyId),
-        ("overmaxrecordsize", S3ErrorCode::OverMaxRecordSize),
-        ("preconditionfailed", S3ErrorCode::PreconditionFailed),
-        ("unrecognizedformatexception", S3ErrorCode::UnrecognizedFormatException),
-        ("parseunsupportedcaseclause", S3ErrorCode::ParseUnsupportedCaseClause),
-        ("tokencodeinvaliderror", S3ErrorCode::TokenCodeInvalidError),
-        ("parseexpectedidentforat", S3ErrorCode::ParseExpectedIdentForAt),
-        ("invalidsecurity", S3ErrorCode::InvalidSecurity),
-        ("nosuchkey", S3ErrorCode::NoSuchKey),
-        ("invalidbucketname", S3ErrorCode::InvalidBucketName),
-        ("maxmessagelengthexceeded", S3ErrorCode::MaxMessageLengthExceeded),
-        ("csvescapingrecorddelimiter", S3ErrorCode::CSVEscapingRecordDelimiter),
-        ("parseexpectedleftparenaftercast", S3ErrorCode::ParseExpectedLeftParenAfterCast),
-        ("toomanyaccesspoints", S3ErrorCode::TooManyAccessPoints),
-        ("methodnotallowed", S3ErrorCode::MethodNotAllowed),
-        ("lexerinvalidliteral", S3ErrorCode::LexerInvalidLiteral),
-        ("endpointnotfound", S3ErrorCode::EndpointNotFound),
-        ("devicenotactiveerror", S3ErrorCode::DeviceNotActiveError),
-        ("connectionclosedbyrequester", S3ErrorCode::ConnectionClosedByRequester),
-        ("evaluatorbindingdoesnotexist", S3ErrorCode::EvaluatorBindingDoesNotExist),
-        (
-            "evaluatortimestampformatpatternduplicatefields",
-            S3ErrorCode::EvaluatorTimestampFormatPatternDuplicateFields,
-        ),
-        ("parseexpectedwhenclause", S3ErrorCode::ParseExpectedWhenClause),
-        ("bucketnotempty", S3ErrorCode::BucketNotEmpty),
-        ("emptyrequestbody", S3ErrorCode::EmptyRequestBody),
-        ("unsupportedstorageclass", S3ErrorCode::UnsupportedStorageClass),
-        ("busy", S3ErrorCode::Busy),
-        ("unsupportedscanrangeinput", S3ErrorCode::UnsupportedScanRangeInput),
-        ("parseexpectedargumentdelimiter", S3ErrorCode::ParseExpectedArgumentDelimiter),
-        (
-            "evaluatorinvalidtimestampformatpatternsymbol",
-            S3ErrorCode::EvaluatorInvalidTimestampFormatPatternSymbol,
-        ),
-        ("invaliddigest", S3ErrorCode::InvalidDigest),
-        ("expressiontoolong", S3ErrorCode::ExpressionTooLong),
-        ("overmaxparquetblocksize", S3ErrorCode::OverMaxParquetBlockSize),
-        ("slowdown", S3ErrorCode::SlowDown),
-        (
-            "parseinvalidcontextforwildcardinselectlist",
-            S3ErrorCode::ParseInvalidContextForWildcardInSelectList,
-        ),
-        (
-            "evaluatorinvalidtimestampformatpatterntoken",
-            S3ErrorCode::EvaluatorInvalidTimestampFormatPatternToken,
-        ),
-        ("integeroverflow", S3ErrorCode::IntegerOverflow),
-        ("csvparsingerror", S3ErrorCode::CSVParsingError),
-        ("nosuchversion", S3ErrorCode::NoSuchVersion),
-        ("invalidaccesspoint", S3ErrorCode::InvalidAccessPoint),
-        (
-            "evaluatorunterminatedtimestampformatpatterntoken",
-            S3ErrorCode::EvaluatorUnterminatedTimestampFormatPatternToken,
-        ),
-        ("invalidencryptionalgorithmerror", S3ErrorCode::InvalidEncryptionAlgorithmError),
-        (
-            "serversideencryptionconfigurationnotfounderror",
-            S3ErrorCode::ServerSideEncryptionConfigurationNotFoundError,
-        ),
-        ("parseexpectedidentforalias", S3ErrorCode::ParseExpectedIdentForAlias),
-        ("illegallocationconstraintexception", S3ErrorCode::IllegalLocationConstraintException),
-        ("invalidstorageclass", S3ErrorCode::InvalidStorageClass),
-        ("invalidsoaprequest", S3ErrorCode::InvalidSOAPRequest),
-        ("invalidpayer", S3ErrorCode::InvalidPayer),
-        ("invalidaccesspointaliaserror", S3ErrorCode::InvalidAccessPointAliasError),
-        ("csvunescapedquote", S3ErrorCode::CSVUnescapedQuote),
-        ("unsupportedsqlstructure", S3ErrorCode::UnsupportedSqlStructure),
-        ("truncatedinput", S3ErrorCode::TruncatedInput),
-        ("permanentredirect", S3ErrorCode::PermanentRedirect),
-        ("parseunsupportedcase", S3ErrorCode::ParseUnsupportedCase),
-        ("notransformationdefined", S3ErrorCode::NoTransformationDefined),
-        ("invalidcompressionformat", S3ErrorCode::InvalidCompressionFormat),
-        ("malformedpostrequest", S3ErrorCode::MalformedPOSTRequest),
-        (
-            "parseexpectedleftparenbuiltinfunctioncall",
-            S3ErrorCode::ParseExpectedLeftParenBuiltinFunctionCall,
-        ),
-        ("credentialsnotsupported", S3ErrorCode::CredentialsNotSupported),
-        ("authorizationqueryparameterserror", S3ErrorCode::AuthorizationQueryParametersError),
-        (
-            "parseexpectedleftparenvalueconstructor",
-            S3ErrorCode::ParseExpectedLeftParenValueConstructor,
-        ),
-        ("jsonparsingerror", S3ErrorCode::JSONParsingError),
-        ("accountproblem", S3ErrorCode::AccountProblem),
-        (
-            "parseexpectedrightparenbuiltinfunctioncall",
-            S3ErrorCode::ParseExpectedRightParenBuiltinFunctionCall,
-        ),
-        ("missingcontentlength", S3ErrorCode::MissingContentLength),
-        ("objectlockconfigurationnotfounderror", S3ErrorCode::ObjectLockConfigurationNotFoundError),
-        ("operationaborted", S3ErrorCode::OperationAborted),
-        ("nosuchupload", S3ErrorCode::NoSuchUpload),
-        ("nosuchlifecycleconfiguration", S3ErrorCode::NoSuchLifecycleConfiguration),
-        ("toomanymultiregionaccesspoints", S3ErrorCode::TooManyMultiRegionAccessPoints),
-        ("parseexpectedkeyword", S3ErrorCode::ParseExpectedKeyword),
-        ("objectserializationconflict", S3ErrorCode::ObjectSerializationConflict),
-        ("missingauthenticationtoken", S3ErrorCode::MissingAuthenticationToken),
-        ("parseexpectedidentforgroupname", S3ErrorCode::ParseExpectedIdentForGroupName),
-        ("tagpolicyexception", S3ErrorCode::TagPolicyException),
-        ("requesttimetooskewed", S3ErrorCode::RequestTimeTooSkewed),
-        ("invalidsignature", S3ErrorCode::InvalidSignature),
-        ("toomanytags", S3ErrorCode::TooManyTags),
-        ("missingrequiredparameter", S3ErrorCode::MissingRequiredParameter),
-        ("nosuchmultiregionaccesspoint", S3ErrorCode::NoSuchMultiRegionAccessPoint),
-        ("maxoperatorsexceeded", S3ErrorCode::MaxOperatorsExceeded),
-        ("entitytoosmall", S3ErrorCode::EntityTooSmall),
-        ("unexpectediperror", S3ErrorCode::UnexpectedIPError),
-        (
-            "replicationconfigurationnotfounderror",
-            S3ErrorCode::ReplicationConfigurationNotFoundError,
-        ),
-        ("notmodified", S3ErrorCode::NotModified),
-        ("parseexpecteddatepart", S3ErrorCode::ParseExpectedDatePart),
-        ("parseinvalidtypeparam", S3ErrorCode::ParseInvalidTypeParam),
-        ("invalidcast", S3ErrorCode::InvalidCast),
-        ("invalidtablealias", S3ErrorCode::InvalidTableAlias),
-        ("requestisnotmultipartcontent", S3ErrorCode::RequestIsNotMultiPartContent),
-        ("unsupportedrangeheader", S3ErrorCode::UnsupportedRangeHeader),
-        ("parseunexpectedtoken", S3ErrorCode::ParseUnexpectedToken),
-        ("incompletebody", S3ErrorCode::IncompleteBody),
-        ("parseinvalidpathcomponent", S3ErrorCode::ParseInvalidPathComponent),
-        ("nosuchasyncrequest", S3ErrorCode::NoSuchAsyncRequest),
-        ("requestheadersectiontoolarge", S3ErrorCode::RequestHeaderSectionTooLarge),
-        ("invalidpolicydocument", S3ErrorCode::InvalidPolicyDocument),
-        ("invalidquotefields", S3ErrorCode::InvalidQuoteFields),
-        ("invalidfileheaderinfo", S3ErrorCode::InvalidFileHeaderInfo),
-        ("maxpostpredatalengthexceedederror", S3ErrorCode::MaxPostPreDataLengthExceededError),
-        ("nosuchcorsconfiguration", S3ErrorCode::NoSuchCORSConfiguration),
-        ("parseexpectedexpression", S3ErrorCode::ParseExpectedExpression),
-        ("invalidlocationconstraint", S3ErrorCode::InvalidLocationConstraint),
-        ("responseinterrupted", S3ErrorCode::ResponseInterrupted),
-        ("bucketalreadyexists", S3ErrorCode::BucketAlreadyExists),
-        ("serviceunavailable", S3ErrorCode::ServiceUnavailable),
-        ("illegalsqlfunctionargument", S3ErrorCode::IllegalSqlFunctionArgument),
-        ("missingsecurityheader", S3ErrorCode::MissingSecurityHeader),
-        ("allaccessdisabled", S3ErrorCode::AllAccessDisabled),
-        ("tokenrefreshrequired", S3ErrorCode::TokenRefreshRequired),
-        ("accessdenied", S3ErrorCode::AccessDenied),
-        ("parseasteriskisnotaloneinselectlist", S3ErrorCode::ParseAsteriskIsNotAloneInSelectList),
-        ("malformedaclerror", S3ErrorCode::MalformedACLError),
-        ("parseexpected2tokentypes", S3ErrorCode::ParseExpected2TokenTypes),
-        ("parseunexpectedoperator", S3ErrorCode::ParseUnexpectedOperator),
-        ("malformedxml", S3ErrorCode::MalformedXML),
-        ("ambiguousfieldname", S3ErrorCode::AmbiguousFieldName),
-        ("crosslocationloggingprohibited", S3ErrorCode::CrossLocationLoggingProhibited),
-        ("redirect", S3ErrorCode::Redirect),
-        ("nologgingstatusforkey", S3ErrorCode::NoLoggingStatusForKey),
-        ("incorrectnumberoffilesinpostrequest", S3ErrorCode::IncorrectNumberOfFilesInPostRequest),
-        (
-            "evaluatorinvalidtimestampformatpatternsymbolforparsing",
-            S3ErrorCode::EvaluatorInvalidTimestampFormatPatternSymbolForParsing,
-        ),
-        ("parsenonunaryagregatefunctioncall", S3ErrorCode::ParseNonUnaryAgregateFunctionCall),
-        ("nosuchbucket", S3ErrorCode::NoSuchBucket),
-        ("missingsecurityelement", S3ErrorCode::MissingSecurityElement),
-        ("invalidargument", S3ErrorCode::InvalidArgument),
-        ("invalidbucketownerawsaccountid", S3ErrorCode::InvalidBucketOwnerAWSAccountID),
-        ("invalidtextencoding", S3ErrorCode::InvalidTextEncoding),
-        ("requesttorrentofbucketerror", S3ErrorCode::RequestTorrentOfBucketError),
-        ("parseexpectednumber", S3ErrorCode::ParseExpectedNumber),
-        (
-            "illegalversioningconfigurationexception",
-            S3ErrorCode::IllegalVersioningConfigurationException,
-        ),
-        ("invalidpartorder", S3ErrorCode::InvalidPartOrder),
-        ("parseselectmissingfrom", S3ErrorCode::ParseSelectMissingFrom),
-        ("externalevalexception", S3ErrorCode::ExternalEvalException),
-        ("invalidrange", S3ErrorCode::InvalidRange),
-        (
-            "evaluatorinvalidtimestampformatpattern",
-            S3ErrorCode::EvaluatorInvalidTimestampFormatPattern,
-        ),
-        ("notimplemented", S3ErrorCode::NotImplemented),
-    ],
-};
-
-static S3_ERROR_CODE_DEFAULT_MESSAGE: phf::Map<&'static str, &'static str> = ::phf::Map {
-    key: 16287231350648472473,
-    disps: &[
-        (0, 4),
-        (0, 1),
-        (0, 10),
-        (0, 2),
-        (0, 2),
-        (0, 0),
-        (0, 0),
-        (0, 18),
-        (0, 53),
-        (0, 119),
-        (0, 27),
-        (0, 0),
-        (0, 12),
-        (0, 0),
-        (0, 1),
-        (0, 33),
-        (0, 100),
-        (1, 0),
-        (0, 37),
-        (0, 15),
-        (0, 1),
-        (0, 1),
-        (0, 1),
-        (0, 67),
-        (0, 11),
-        (0, 0),
-        (0, 10),
-        (1, 0),
-        (0, 14),
-        (0, 73),
-        (0, 9),
-        (0, 92),
-        (0, 2),
-        (0, 3),
-        (0, 2),
-        (0, 4),
-        (0, 48),
-        (0, 35),
-        (0, 11),
-        (0, 83),
-        (0, 48),
-        (0, 71),
-        (0, 10),
-        (0, 20),
-        (0, 82),
-        (0, 17),
-        (0, 29),
-        (0, 22),
-        (0, 4),
-        (0, 1),
-        (0, 111),
-        (0, 47),
-        (0, 99),
-        (0, 30),
-        (0, 39),
-        (0, 103),
-        (0, 32),
-        (1, 123),
-        (0, 209),
-        (0, 7),
-        (0, 9),
-        (0, 0),
-        (0, 54),
-        (0, 0),
-        (0, 43),
-        (0, 3),
-        (0, 1),
-        (0, 40),
-        (0, 18),
-        (0, 17),
-        (0, 79),
-        (0, 3),
-        (0, 46),
-        (0, 114),
-        (0, 2),
-        (0, 103),
-        (0, 2),
-        (0, 114),
-        (0, 112),
-        (0, 3),
-    ],
-    entries: &[
-        ("ParseUnExpectedKeyword", "The SQL expression contains an unexpected keyword."),
-        ("MaxMessageLengthExceeded", "Your request was too big."),
-        (
-            "CSVParsingError",
-            "An error occurred while parsing the CSV file. Check the file and try again.",
-        ),
-        ("EmptyRequestBody", "The request body cannot be empty."),
-        ("AccessControlListNotSupported", "The bucket does not allow ACLs."),
-        ("MalformedPolicy", "Your policy contains a principal that is not valid."),
-        (
-            "ParseAsteriskIsNotAloneInSelectList",
-            "Other expressions are not allowed in the SELECT list when * is used without dot notation in the SQL expression.",
-        ),
-        ("InvalidAddressingHeader", "You must specify the Anonymous role."),
-        (
-            "RequestHeaderSectionTooLarge",
-            "The request header and query parameters used to make the request exceed the maximum allowed size.",
-        ),
-        (
-            "MalformedPOSTRequest",
-            "The body of your POST request is not well-formed multipart/form-data.",
-        ),
-        (
-            "OperationAborted",
-            "A conflicting conditional action is currently in progress against this resource. Try again.",
-        ),
-        ("AuthorizationHeaderMalformed", "The authorization header you provided is invalid."),
-        (
-            "InvalidEncryptionAlgorithmError",
-            "The encryption request you specified is not valid. The valid value is AES256.",
-        ),
-        ("IllegalLocationConstraintException", "The specified location constraint is not valid."),
-        ("InvalidStorageClass", "The storage class you specified is not valid."),
-        ("InvalidObjectState", "The action is not valid for the current state of the object."),
-        ("UnsupportedFunction", "We encountered an unsupported SQL function."),
-        ("UnsupportedRangeHeader", "A range header is not supported for this operation."),
-        ("ParseExpectedTypeName", "The expected type name in the SQL expression was not found."),
-        (
-            "PermanentRedirectControlError",
-            "The API operation you are attempting to access must be addressed using the specified endpoint. Send all future requests to this endpoint.",
-        ),
-        (
-            "CSVEscapingRecordDelimiter",
-            "A quoted record delimiter was found in the file. To allow quoted record delimiters, set AllowQuotedRecordDelimiter to 'TRUE'.",
-        ),
-        (
-            "AmbiguousFieldName",
-            "The field name matches to multiple fields in the file. Check the SQL expression and the file, and try again.",
-        ),
-        ("InvalidBucketName", "The specified bucket is not valid."),
-        (
-            "EvaluatorInvalidTimestampFormatPattern",
-            "The timestamp format string in the SQL expression is not valid.",
-        ),
-        ("EntityTooSmall", "Your proposed upload is smaller than the minimum allowed object size."),
-        ("InvalidURI", "Couldn't parse the specified URI."),
-        ("MissingSecurityHeader", "Your request is missing a required header."),
-        ("InvalidTableAlias", "The SQL expression contains a table alias that is not valid."),
-        (
-            "UnsupportedScanRangeInput",
-            "Scan range queries are not supported on this type of object.",
-        ),
-        (
-            "PermanentRedirect",
-            "The bucket you are attempting to access must be addressed using the specified endpoint. Send all future requests to this endpoint.",
-        ),
-        (
-            "AllAccessDisabled",
-            "All access to this Amazon S3 resource has been disabled. Contact AWS Support for further assistance.",
-        ),
-        ("UnsupportedSqlOperation", "We encountered an unsupported SQL operation."),
-        ("NotImplemented", "A header you provided implies functionality that is not implemented."),
-        (
-            "InvalidPayer",
-            "All access to this object has been disabled. Please contact AWS Support for further assistance.",
-        ),
-        ("InvalidDataType", "The SQL expression contains a data type that is not valid."),
-        (
-            "InvalidPolicyDocument",
-            "The content of the form does not meet the conditions specified in the policy document.",
-        ),
-        (
-            "IncorrectEndpoint",
-            "The specified bucket exists in another Region. Direct requests to the correct endpoint.",
-        ),
-        (
-            "ParseExpectedArgumentDelimiter",
-            "The expected argument delimiter in the SQL expression was not found.",
-        ),
-        ("LexerInvalidChar", "The SQL expression contains a character that is not valid."),
-        (
-            "EvaluatorInvalidTimestampFormatPatternSymbolForParsing",
-            "The timestamp format pattern contains a valid format symbol that cannot be applied to timestamp parsing in the SQL expression.",
-        ),
-        (
-            "ConnectionClosedByRequester",
-            "Returned to the original caller when an error is encountered while reading the WriteGetObjectResponse body.",
-        ),
-        (
-            "InvalidDataSource",
-            "The data source type is not valid. Only CSV, JSON, and Parquet are supported.",
-        ),
-        (
-            "EvaluatorInvalidTimestampFormatPatternSymbol",
-            "The timestamp format pattern contains a symbol in the SQL expression that is not valid.",
-        ),
-        (
-            "InvalidJsonType",
-            "The JsonType value is not valid. Only DOCUMENT and LINES are supported.",
-        ),
-        ("NoSuchLifecycleConfiguration", "The lifecycle configuration does not exist."),
-        (
-            "IncompleteBody",
-            "You did not provide the number of bytes specified by the Content-Length HTTP header",
-        ),
-        ("ParseUnexpectedToken", "The SQL expression contains an unexpected token."),
-        (
-            "UnsupportedTypeForQuerying",
-            "Your query contains an unsupported type for comparison (e.g. verifying that a Parquet INT96 column type is greater than 0).",
-        ),
-        (
-            "InvalidQuoteFields",
-            "The QuoteFields value is not valid. Only ALWAYS and ASNEEDED are supported.",
-        ),
-        (
-            "UserKeyMustBeSpecified",
-            "The bucket POST must contain the specified field name. If it is specified, check the order of the fields.",
-        ),
-        (
-            "ClientTokenConflict",
-            "Your Multi-Region Access Point idempotency token was already used for a different request.",
-        ),
-        ("EndpointNotFound", "Direct requests to the correct endpoint."),
-        ("OwnershipControlsNotFoundError", "The bucket ownership controls were not found."),
-        ("ExpiredToken", "The provided token has expired."),
-        ("ParseUnsupportedSyntax", "The SQL expression contains unsupported syntax."),
-        ("InvalidArgument", "Invalid argument."),
-        (
-            "MaxOperatorsExceeded",
-            "Failed to parse SQL expression, try reducing complexity. For example, reduce number of operators used.",
-        ),
-        (
-            "CastFailed",
-            "An attempt to convert from one data type to another using CAST failed in the SQL expression.",
-        ),
-        ("ExternalEvalException", "The query cannot be evaluated. Check the file and try again."),
-        ("NoSuchBucketPolicy", "The specified bucket does not have a bucket policy."),
-        (
-            "InvalidHostHeader",
-            "The host headers provided in the request used the incorrect style addressing.",
-        ),
-        ("LexerInvalidLiteral", "The SQL expression contains an operator that is not valid."),
-        (
-            "InvalidSignature",
-            "The request signature that the server calculated does not match the signature that you provided. Check your AWS secret access key and signing method. For more information, see Signing and authenticating REST requests.",
-        ),
-        ("ParseUnsupportedSelect", "The SQL expression contains an unsupported use of SELECT."),
-        ("ParseExpectedDatePart", "The expected date part in the SQL expression was not found."),
-        ("InvalidBucketState", "The request is not valid with the current state of the bucket."),
-        (
-            "TruncatedInput",
-            "Object decompression failed. Check that the object is properly compressed using the format specified in the request.",
-        ),
-        ("MissingContentLength", "You must provide the Content-Length HTTP header."),
-        (
-            "ParseInvalidPathComponent",
-            "The SQL expression contains a path component that is not valid.",
-        ),
-        ("ParseExpectedIdentForGroupName", "GROUP is not supported in the SQL expression."),
-        ("DeviceNotActiveError", "The device is not currently active."),
-        (
-            "Redirect",
-            "Temporary redirect. You are being redirected to the bucket while the Domain Name System (DNS) server is being updated.",
-        ),
-        ("LexerInvalidOperator", "The SQL expression contains a literal that is not valid."),
-        (
-            "InvalidTargetBucketForLogging",
-            "The target bucket for logging does not exist, is not owned by you, or does not have the appropriate grants for the log-delivery group.",
-        ),
-        ("MethodNotAllowed", "The specified method is not allowed against this resource."),
-        ("MissingAttachment", "A SOAP attachment was expected, but none were found."),
-        ("InvalidRequest", "Invalid request."),
-        ("InvalidSecurity", "The provided security credentials are not valid."),
-        (
-            "ExpressionTooLong",
-            "The SQL expression is too long. The maximum byte-length for an SQL expression is 256 KB.",
-        ),
-        (
-            "ParseNonUnaryAgregateFunctionCall",
-            "Only one argument is supported for aggregate functions in the SQL expression.",
-        ),
-        (
-            "ParseInvalidTypeParam",
-            "The SQL expression contains a parameter value that is not valid.",
-        ),
-        (
-            "MalformedXML",
-            "This happens when the user sends malformed XML (XML that doesn't conform to the published XSD) for the configuration. The error message is, \"The XML you provided was not well-formed or did not validate against our published schema.\"",
-        ),
-        ("ParseExpectedKeyword", "The expected keyword in the SQL expression was not found."),
-        ("RequestTorrentOfBucketError", "Requesting the torrent file of a bucket is not permitted."),
-        (
-            "UnexpectedIPError",
-            "Applicable in China Regions only. This request was rejected because the IP was unexpected.",
-        ),
-        (
-            "InvalidCast",
-            "An attempt to convert from one data type to another using CAST failed in the SQL expression.",
-        ),
-        (
-            "ParseExpectedRightParenBuiltinFunctionCall",
-            "The expected right parenthesis character in the SQL expression was not found.",
-        ),
-        ("ParseExpectedTokenType", "The expected token in the SQL expression was not found."),
-        (
-            "AccountProblem",
-            "There is a problem with your AWS account that prevents the action from completing successfully. Contact AWS Support for further assistance.",
-        ),
-        ("AccessDenied", "Access Denied"),
-        ("InvalidAccessPointAliasError", "The specified access point alias name is not valid."),
-        ("InvalidColumnIndex", "The column index in the SQL expression is not valid."),
-        ("InvalidDigest", "The Content-MD5 you specified is not valid."),
-        (
-            "NotSignedUp",
-            "Your account is not signed up for the Amazon S3 service. You must sign up before you can use Amazon S3. You can sign up at the following URL: Amazon S3",
-        ),
-        (
-            "ParseExpectedIdentForAlias",
-            "The expected identifier for the alias in the SQL expression was not found.",
-        ),
-        (
-            "ObjectSerializationConflict",
-            "InputSerialization specifies more than one format (CSV, JSON, or Parquet), or OutputSerialization specifies more than one format (CSV or JSON). For InputSerialization and OutputSerialization, you can specify only one format for each.",
-        ),
-        ("ParseUnexpectedOperator", "The SQL expression contains an unexpected operator."),
-        ("TokenCodeInvalidError", "The serial number and/or token code you provided is not valid."),
-        ("ParseExpectedExpression", "The expected SQL expression was not found."),
-        ("ParseExpectedMember", "The SQL expression contains an unsupported use of MEMBER."),
-        ("PreconditionFailed", "At least one of the preconditions you specified did not hold."),
-        ("BadDigest", "The Content-MD5 you specified did not match what we received."),
-        (
-            "InvalidPart",
-            "One or more of the specified parts could not be found. The part might not have been uploaded, or the specified entity tag might not have matched the part's entity tag.",
-        ),
-        (
-            "InvalidLocationConstraint",
-            "The specified location constraint is not valid. For more information about Regions, see How to Select a Region for Your Buckets.",
-        ),
-        (
-            "ParquetUnsupportedCompressionCodec",
-            "The specified Parquet compression codec is not supported.",
-        ),
-        ("ParseCastArity", "The SQL expression CAST has incorrect arity."),
-        ("NoTransformationDefined", "No transformation found for this Object Lambda Access Point."),
-        (
-            "InvalidCompressionFormat",
-            "The file is not in a supported compression format. Only GZIP and BZIP2 are supported.",
-        ),
-        ("NoSuchAccessPoint", "The specified access point does not exist."),
-        ("InlineDataTooLarge", "Inline data exceeds the maximum allowed size."),
-        (
-            "TooManyMultiRegionAccessPoints",
-            "You have attempted to create more Multi-Region Access Points than are allowed for an account. For more information, see Amazon Simple Storage Service endpoints and quotas in the AWS General Reference.",
-        ),
-        (
-            "ParseUnsupportedCallWithStar",
-            "Only COUNT with (*) as a parameter is supported in the SQL expression.",
-        ),
-        ("ParseUnexpectedTerm", "The SQL expression contains an unexpected term."),
-        ("Busy", "The service is unavailable. Try again later."),
-        ("InvalidTextEncoding", "The encoding type is not valid. Only UTF-8 encoding is supported."),
-        (
-            "IncorrectSqlFunctionArgumentType",
-            "An incorrect argument type was specified in a function call in the SQL expression.",
-        ),
-        (
-            "UnresolvableGrantByEmailAddress",
-            "The email address you provided does not match any account on record.",
-        ),
-        (
-            "NoSuchVersion",
-            "Indicates that the version ID specified in the request does not match an existing version.",
-        ),
-        ("InvalidRange", "The requested range cannot be satisfied."),
-        (
-            "InvalidPartOrder",
-            "The list of parts was not in ascending order. Parts list must be specified in order by part number.",
-        ),
-        ("BucketNotEmpty", "The bucket you tried to delete is not empty."),
-        ("TemporaryRedirect", "You are being redirected to the bucket while DNS updates."),
-        ("MissingSecurityElement", "The SOAP 1.1 request is missing a security element."),
-        (
-            "TooManyAccessPoints",
-            "You have attempted to create more access points than are allowed for an account. For more information, see Amazon Simple Storage Service endpoints and quotas in the AWS General Reference.",
-        ),
-        (
-            "UnauthorizedAccessError",
-            "Applicable in China Regions only. Returned when a request is made to a bucket that doesn't have an ICP license. For more information, see ICP Recordal.",
-        ),
-        (
-            "NoSuchWebsiteConfiguration",
-            "The specified bucket does not have a website configuration.",
-        ),
-        (
-            "ParseExpectedWhenClause",
-            "The expected WHEN clause in the SQL expression was not found. CASE is not supported.",
-        ),
-        ("RestoreAlreadyInProgress", "Object restore is already in progress."),
-        ("MissingAuthenticationToken", "The request was not signed."),
-        ("ParseExpected2TokenTypes", "The expected token in the SQL expression was not found."),
-        (
-            "AccessPointAlreadyOwnedByYou",
-            "An access point with an identical name already exists in your account.",
-        ),
-        (
-            "UnsupportedSignature",
-            "The provided request is signed with an unsupported STS Token version or the signature version is not supported.",
-        ),
-        ("ParseUnsupportedCaseClause", "The SQL expression contains an unsupported use of CASE."),
-        (
-            "TooManyMultiRegionAccessPointregionsError",
-            "You have attempted to create a Multi-Region Access Point with more Regions than are allowed for an account. For more information, see Amazon Simple Storage Service endpoints and quotas in the AWS General Reference.",
-        ),
-        (
-            "NumberFormatError",
-            "An error occurred while parsing a number. This error can be caused by underflow or overflow of integers.",
-        ),
-        (
-            "ParseExpectedLeftParenValueConstructor",
-            "The expected left parenthesis in the SQL expression was not found.",
-        ),
-        (
-            "MissingRequiredParameter",
-            "The SelectRequest entity is missing a required parameter. Check the service documentation and try again.",
-        ),
-        (
-            "NoSuchObjectLockConfiguration",
-            "The specified object does not have an ObjectLock configuration.",
-        ),
-        ("NoSuchCORSConfiguration", "The specified bucket does not have a CORS configuration."),
-        (
-            "ReplicationConfigurationNotFoundError",
-            "There is no replication configuration for this bucket.",
-        ),
-        ("ParseEmptySelect", "The SQL expression contains an empty SELECT clause."),
-        (
-            "AuthorizationQueryParametersError",
-            "The authorization query parameters that you provided are not valid.",
-        ),
-        (
-            "SignatureDoesNotMatch",
-            "The request signature we calculated does not match the signature you provided. Check your AWS secret access key and signing method. For more information, see REST Authentication and SOAP Authentication for details.",
-        ),
-        ("TooManyTags", "The number of tags exceeds the limit of 50 tags."),
-        (
-            "ColumnTooLong",
-            "The length of a column in the result is greater than maxCharsPerColumn of 1 MB.",
-        ),
-        (
-            "JSONParsingError",
-            "An error occurred while parsing the JSON file. Check the file and try again.",
-        ),
-        (
-            "BucketAlreadyOwnedByYou",
-            "The bucket you tried to create already exists, and you own it. Amazon S3 returns this error in all AWS Regions except in the North Virginia Region. For legacy compatibility, if you re-create an existing bucket that you already own in the North Virginia Region, Amazon S3 returns 200 OK and resets the bucket access control lists (ACLs).",
-        ),
-        (
-            "BucketAlreadyExists",
-            "The requested bucket name is not available. The bucket namespace is shared by all users of the system. Please select a different name and try again.",
-        ),
-        (
-            "AmbiguousGrantByEmailAddress",
-            "The email address you provided is associated with more than one account.",
-        ),
-        (
-            "ServerSideEncryptionConfigurationNotFoundError",
-            "The server-side encryption configuration was not found.",
-        ),
-        (
-            "InvalidFileHeaderInfo",
-            "The FileHeaderInfo value is not valid. Only NONE, USE, and IGNORE are supported.",
-        ),
-        (
-            "TooManyBuckets",
-            "You have attempted to create more buckets than are allowed for an account. For more information, see Amazon Simple Storage Service endpoints and quotas in the AWS General Reference.",
-        ),
-        (
-            "ObjectLockConfigurationNotFoundError",
-            "The Object Lock configuration does not exist for this bucket.",
-        ),
-        (
-            "ResponseInterrupted",
-            "Returned to the original caller when an error is encountered while reading the WriteGetObjectResponse body.",
-        ),
-        ("NoSuchMultiRegionAccessPoint", "The specified Multi-Region Access Point does not exist."),
-        (
-            "BucketHasAccessPointsAttached",
-            "The bucket you tried to delete has access points attached. Delete your access points before deleting your bucket.",
-        ),
-        (
-            "EvaluatorTimestampFormatPatternDuplicateFields",
-            "The timestamp format pattern contains multiple format specifiers representing the timestamp field in the SQL expression.",
-        ),
-        (
-            "InvalidBucketOwnerAWSAccountID",
-            "The value of the expected bucket owner parameter must be an AWS account ID.",
-        ),
-        (
-            "IllegalVersioningConfigurationException",
-            "Indicates that the versioning configuration specified in the request is invalid.",
-        ),
-        (
-            "EvaluatorBindingDoesNotExist",
-            "A column name or a path provided does not exist in the SQL expression.",
-        ),
-        (
-            "MalformedACLError",
-            "The XML you provided was not well-formed or did not validate against our published schema.",
-        ),
-        (
-            "ParseCannotMixSqbAndWildcardInSelectList",
-            "Cannot mix [] and * in the same expression in a SELECT list in the SQL expression.",
-        ),
-        ("UnsupportedParquetType", "The specified Parquet type is not supported."),
-        ("ParseUnsupportedToken", "The SQL expression contains an unsupported token."),
-        (
-            "InvalidTag",
-            "Your request contains tag input that is not valid. For example, your request might contain duplicate keys, keys or values that are too long, or system tags.",
-        ),
-        (
-            "UnsupportedSqlStructure",
-            "We encountered an unsupported SQL structure. Check the SQL Reference.",
-        ),
-        (
-            "NotDeviceOwnerError",
-            "The device that generated the token is not owned by the authenticated user.",
-        ),
-        ("CredentialsNotSupported", "This request does not support credentials."),
-        (
-            "EvaluatorInvalidArguments",
-            "There is an incorrect number of arguments in the function call in the SQL expression.",
-        ),
-        ("NoSuchBucket", "The specified bucket does not exist."),
-        (
-            "TagPolicyException",
-            "The tag policy does not allow the specified value for the following tag key.",
-        ),
-        (
-            "EvaluatorTimestampFormatPatternHourClockAmPmMismatch",
-            "The timestamp format pattern contains a 12-hour hour of day format symbol but doesn't also contain an AM/PM field, or it contains a 24-hour hour of day format specifier and contains an AM/PM field in the SQL expression.",
-        ),
-        ("InvalidScanRange", "The provided scan range is not valid."),
-        (
-            "ParseMissingIdentAfterAt",
-            "The expected identifier after the @ symbol in the SQL expression was not found.",
-        ),
-        ("MetadataTooLarge", "Your metadata headers exceed the maximum allowed metadata size."),
-        (
-            "EvaluatorInvalidTimestampFormatPatternToken",
-            "The timestamp format pattern contains a token in the SQL expression that is not valid.",
-        ),
-        ("ParseMalformedJoin", "JOIN is not supported in the SQL expression."),
-        ("ParseExpectedNumber", "The expected number in the SQL expression was not found."),
-        ("SlowDown", "Reduce your request rate."),
-        (
-            "InvalidExpressionType",
-            "The ExpressionType value is not valid. Only SQL expressions are supported.",
-        ),
-        (
-            "ParseExpectedLeftParenBuiltinFunctionCall",
-            "The expected left parenthesis in the SQL expression was not found.",
-        ),
-        ("ValueParseFailure", "A timestamp parse failure occurred in the SQL expression."),
-        ("ParseUnsupportedCase", "The SQL expression contains an unsupported use of CASE."),
-        (
-            "ParseExpectedIdentForAt",
-            "The expected identifier for AT name in the SQL expression was not found.",
-        ),
-        ("InvalidAccessKeyId", "The AWS access key ID you provided does not exist in our records."),
-        (
-            "EvaluatorUnterminatedTimestampFormatPatternToken",
-            "The timestamp format pattern contains an unterminated token in the SQL expression.",
-        ),
-        ("NoSuchAsyncRequest", "The specified request was not found."),
-        ("NotModified", "The resource was not changed."),
-        ("ServiceUnavailable", "Service is unable to handle request."),
-        ("IllegalSqlFunctionArgument", "An illegal argument was used in the SQL function."),
-        ("UnsupportedArgument", "The request contained an unsupported argument."),
-        ("UnauthorizedAccess", "You are not authorized to perform this operation."),
-        (
-            "CSVUnescapedQuote",
-            "An unescaped quote was found while parsing the CSV file. To allow quoted record delimiters, set AllowQuotedRecordDelimiter to 'TRUE'.",
-        ),
-        ("TokenRefreshRequired", "The provided token must be refreshed."),
-        (
-            "CrossLocationLoggingProhibited",
-            "Cross-location logging not allowed. Buckets in one geographic location cannot log information to a bucket in another location.",
-        ),
-        (
-            "ParseInvalidContextForWildcardInSelectList",
-            "The use of * in the SELECT list in the SQL expression is not valid.",
-        ),
-        ("InvalidAccessPoint", "The specified access point name or account is not valid."),
-        (
-            "UnsupportedStorageClass",
-            "We encountered a storage class that is not supported. Only STANDARD, STANDARD_IA, and ONEZONE_IA storage classes are supported.",
-        ),
-        (
-            "NoSuchUpload",
-            "The specified multipart upload does not exist. The upload ID might be invalid, or the multipart upload might have been aborted or completed.",
-        ),
-        (
-            "RequestIsNotMultiPartContent",
-            "Bucket POST must be of the enclosure-type multipart/form-data.",
-        ),
-        (
-            "InvalidSessionException",
-            "Returned if the session doesn't exist anymore because it timed out or expired.",
-        ),
-        (
-            "EvaluatorLikePatternInvalidEscapeSequence",
-            "An argument given to the LIKE expression was not valid.",
-        ),
-        ("UnexpectedContent", "This request does not support content."),
-        (
-            "NoLoggingStatusForKey",
-            "There is no such thing as a logging status subresource for a key.",
-        ),
-        ("ParseUnknownOperator", "The SQL expression contains an operator that is not valid."),
-        ("EvaluatorNegativeLimit", "LIMIT must not be negative."),
-        (
-            "OverMaxColumn",
-            "The number of columns in the result is greater than the maximum allowable number of columns.",
-        ),
-        (
-            "ConditionalRequestConflict",
-            "A conflicting operation occurred. If using PutObject you can retry the request. If using multipart upload you should initiate another CreateMultipartUpload request and re-upload each part.",
-        ),
-        (
-            "InvalidBucketAclWithObjectOwnership",
-            "Bucket cannot have ACLs set with ObjectOwnership's BucketOwnerEnforced setting.",
-        ),
-        (
-            "InvalidRegion",
-            "You've attempted to create a Multi-Region Access Point in a Region that you haven't opted in to.",
-        ),
-        ("InvalidSOAPRequest", "The SOAP request body is invalid."),
-        ("InvalidKeyPath", "The key path in the SQL expression is not valid."),
-        ("InternalError", "We encountered an internal error. Please try again."),
-        ("OverMaxParquetBlockSize", "The Parquet file is above the max row group size."),
-        ("UnrecognizedFormatException", "We encountered a record type that is not valid."),
-        ("LexerInvalidIONLiteral", "The SQL expression contains an operator that is not valid."),
-        ("InvalidHttpMethod", "The request is made using an unexpected HTTP method."),
-        (
-            "ParquetParsingError",
-            "An error occurred while parsing the Parquet file. Check the file and try again.",
-        ),
-        (
-            "IncorrectNumberOfFilesInPostRequest",
-            "POST requires exactly one file upload per request.",
-        ),
-        (
-            "ParseExpectedLeftParenAfterCast",
-            "The expected left parenthesis after CAST in the SQL expression was not found.",
-        ),
-        ("KeyTooLongError", "Your key is too long."),
-        (
-            "OverMaxRecordSize",
-            "The length of a record in the input or result is greater than the maxCharsPerRecord limit of 1 MB.",
-        ),
-        (
-            "RequestTimeout",
-            "Your socket connection to the server was not read from or written to within the timeout period.",
-        ),
-        (
-            "LikeInvalidInputs",
-            "The argument given to the LIKE clause in the SQL expression is not valid.",
-        ),
-        ("InvalidToken", "The provided token is malformed or otherwise invalid."),
-        (
-            "ParseUnsupportedLiteralsGroupBy",
-            "The SQL expression contains an unsupported use of GROUP BY.",
-        ),
-        ("IntegerOverflow", "An integer overflow or underflow occurred in the SQL expression."),
-        ("UnsupportedSyntax", "We encountered syntax that is not valid."),
-        (
-            "MaxPostPreDataLengthExceededError",
-            "Your POST request fields preceding the upload file were too large.",
-        ),
-        (
-            "RequestTimeTooSkewed",
-            "The difference between the request time and the server's time is too large.",
-        ),
-        (
-            "InvalidRequestParameter",
-            "The value of a parameter in the SelectRequest element is not valid. Check the service API documentation and try again.",
-        ),
-        ("NoSuchTagSet", "The specified tag does not exist."),
-        (
-            "MissingRequestBodyError",
-            "This happens when the user sends an empty XML document as a request. The error message is, \"Request body is empty.\"",
-        ),
-        ("EntityTooLarge", "Your proposed upload exceeds the maximum allowed object size."),
-        ("MultipleDataSourcesUnsupported", "Multiple data sources are not supported."),
-        ("NoSuchResource", "The specified resource doesn't exist."),
-        ("NoSuchKey", "The specified key does not exist."),
-        ("ParseUnsupportedAlias", "The SQL expression contains an unsupported use of ALIAS."),
-        (
-            "ParseSelectMissingFrom",
-            "The SQL expression contains a missing FROM after the SELECT list.",
-        ),
-    ],
-};
-
-impl S3ErrorCode {
-    pub(super) const STATIC_CODE_LIST: &'static [&'static str] = &[
-        "AccessControlListNotSupported",
-        "AccessDenied",
-        "AccessPointAlreadyOwnedByYou",
-        "AccountProblem",
-        "AllAccessDisabled",
-        "AmbiguousFieldName",
-        "AmbiguousGrantByEmailAddress",
-        "AuthorizationHeaderMalformed",
-        "AuthorizationQueryParametersError",
-        "BadDigest",
-        "BucketAlreadyExists",
-        "BucketAlreadyOwnedByYou",
-        "BucketHasAccessPointsAttached",
-        "BucketNotEmpty",
-        "Busy",
-        "CSVEscapingRecordDelimiter",
-        "CSVParsingError",
-        "CSVUnescapedQuote",
-        "CastFailed",
-        "ClientTokenConflict",
-        "ColumnTooLong",
-        "ConditionalRequestConflict",
-        "ConnectionClosedByRequester",
-        "CredentialsNotSupported",
-        "CrossLocationLoggingProhibited",
-        "DeviceNotActiveError",
-        "EmptyRequestBody",
-        "EndpointNotFound",
-        "EntityTooLarge",
-        "EntityTooSmall",
-        "EvaluatorBindingDoesNotExist",
-        "EvaluatorInvalidArguments",
-        "EvaluatorInvalidTimestampFormatPattern",
-        "EvaluatorInvalidTimestampFormatPatternSymbol",
-        "EvaluatorInvalidTimestampFormatPatternSymbolForParsing",
-        "EvaluatorInvalidTimestampFormatPatternToken",
-        "EvaluatorLikePatternInvalidEscapeSequence",
-        "EvaluatorNegativeLimit",
-        "EvaluatorTimestampFormatPatternDuplicateFields",
-        "EvaluatorTimestampFormatPatternHourClockAmPmMismatch",
-        "EvaluatorUnterminatedTimestampFormatPatternToken",
-        "ExpiredToken",
-        "ExpressionTooLong",
-        "ExternalEvalException",
-        "IllegalLocationConstraintException",
-        "IllegalSqlFunctionArgument",
-        "IllegalVersioningConfigurationException",
-        "IncompleteBody",
-        "IncorrectEndpoint",
-        "IncorrectNumberOfFilesInPostRequest",
-        "IncorrectSqlFunctionArgumentType",
-        "InlineDataTooLarge",
-        "IntegerOverflow",
-        "InternalError",
-        "InvalidAccessKeyId",
-        "InvalidAccessPoint",
-        "InvalidAccessPointAliasError",
-        "InvalidAddressingHeader",
-        "InvalidArgument",
-        "InvalidBucketAclWithObjectOwnership",
-        "InvalidBucketName",
-        "InvalidBucketOwnerAWSAccountID",
-        "InvalidBucketState",
-        "InvalidCast",
-        "InvalidColumnIndex",
-        "InvalidCompressionFormat",
-        "InvalidDataSource",
-        "InvalidDataType",
-        "InvalidDigest",
-        "InvalidEncryptionAlgorithmError",
-        "InvalidExpressionType",
-        "InvalidFileHeaderInfo",
-        "InvalidHostHeader",
-        "InvalidHttpMethod",
-        "InvalidJsonType",
-        "InvalidKeyPath",
-        "InvalidLocationConstraint",
-        "InvalidObjectState",
-        "InvalidPart",
-        "InvalidPartOrder",
-        "InvalidPayer",
-        "InvalidPolicyDocument",
-        "InvalidQuoteFields",
-        "InvalidRange",
-        "InvalidRegion",
-        "InvalidRequest",
-        "InvalidRequestParameter",
-        "InvalidSOAPRequest",
-        "InvalidScanRange",
-        "InvalidSecurity",
-        "InvalidSessionException",
-        "InvalidSignature",
-        "InvalidStorageClass",
-        "InvalidTableAlias",
-        "InvalidTag",
-        "InvalidTargetBucketForLogging",
-        "InvalidTextEncoding",
-        "InvalidToken",
-        "InvalidURI",
-        "JSONParsingError",
-        "KeyTooLongError",
-        "LexerInvalidChar",
-        "LexerInvalidIONLiteral",
-        "LexerInvalidLiteral",
-        "LexerInvalidOperator",
-        "LikeInvalidInputs",
-        "MalformedACLError",
-        "MalformedPOSTRequest",
-        "MalformedPolicy",
-        "MalformedXML",
-        "MaxMessageLengthExceeded",
-        "MaxOperatorsExceeded",
-        "MaxPostPreDataLengthExceededError",
-        "MetadataTooLarge",
-        "MethodNotAllowed",
-        "MissingAttachment",
-        "MissingAuthenticationToken",
-        "MissingContentLength",
-        "MissingRequestBodyError",
-        "MissingRequiredParameter",
-        "MissingSecurityElement",
-        "MissingSecurityHeader",
-        "MultipleDataSourcesUnsupported",
-        "NoLoggingStatusForKey",
-        "NoSuchAccessPoint",
-        "NoSuchAsyncRequest",
-        "NoSuchBucket",
-        "NoSuchBucketPolicy",
-        "NoSuchCORSConfiguration",
-        "NoSuchKey",
-        "NoSuchLifecycleConfiguration",
-        "NoSuchMultiRegionAccessPoint",
-        "NoSuchObjectLockConfiguration",
-        "NoSuchResource",
-        "NoSuchTagSet",
-        "NoSuchUpload",
-        "NoSuchVersion",
-        "NoSuchWebsiteConfiguration",
-        "NoTransformationDefined",
-        "NotDeviceOwnerError",
-        "NotImplemented",
-        "NotModified",
-        "NotSignedUp",
-        "NumberFormatError",
-        "ObjectLockConfigurationNotFoundError",
-        "ObjectSerializationConflict",
-        "OperationAborted",
-        "OverMaxColumn",
-        "OverMaxParquetBlockSize",
-        "OverMaxRecordSize",
-        "OwnershipControlsNotFoundError",
-        "ParquetParsingError",
-        "ParquetUnsupportedCompressionCodec",
-        "ParseAsteriskIsNotAloneInSelectList",
-        "ParseCannotMixSqbAndWildcardInSelectList",
-        "ParseCastArity",
-        "ParseEmptySelect",
-        "ParseExpected2TokenTypes",
-        "ParseExpectedArgumentDelimiter",
-        "ParseExpectedDatePart",
-        "ParseExpectedExpression",
-        "ParseExpectedIdentForAlias",
-        "ParseExpectedIdentForAt",
-        "ParseExpectedIdentForGroupName",
-        "ParseExpectedKeyword",
-        "ParseExpectedLeftParenAfterCast",
-        "ParseExpectedLeftParenBuiltinFunctionCall",
-        "ParseExpectedLeftParenValueConstructor",
-        "ParseExpectedMember",
-        "ParseExpectedNumber",
-        "ParseExpectedRightParenBuiltinFunctionCall",
-        "ParseExpectedTokenType",
-        "ParseExpectedTypeName",
-        "ParseExpectedWhenClause",
-        "ParseInvalidContextForWildcardInSelectList",
-        "ParseInvalidPathComponent",
-        "ParseInvalidTypeParam",
-        "ParseMalformedJoin",
-        "ParseMissingIdentAfterAt",
-        "ParseNonUnaryAgregateFunctionCall",
-        "ParseSelectMissingFrom",
-        "ParseUnExpectedKeyword",
-        "ParseUnexpectedOperator",
-        "ParseUnexpectedTerm",
-        "ParseUnexpectedToken",
-        "ParseUnknownOperator",
-        "ParseUnsupportedAlias",
-        "ParseUnsupportedCallWithStar",
-        "ParseUnsupportedCase",
-        "ParseUnsupportedCaseClause",
-        "ParseUnsupportedLiteralsGroupBy",
-        "ParseUnsupportedSelect",
-        "ParseUnsupportedSyntax",
-        "ParseUnsupportedToken",
-        "PermanentRedirect",
-        "PermanentRedirectControlError",
-        "PreconditionFailed",
-        "Redirect",
-        "ReplicationConfigurationNotFoundError",
-        "RequestHeaderSectionTooLarge",
-        "RequestIsNotMultiPartContent",
-        "RequestTimeTooSkewed",
-        "RequestTimeout",
-        "RequestTorrentOfBucketError",
-        "ResponseInterrupted",
-        "RestoreAlreadyInProgress",
-        "ServerSideEncryptionConfigurationNotFoundError",
-        "ServiceUnavailable",
-        "SignatureDoesNotMatch",
-        "SlowDown",
-        "TagPolicyException",
-        "TemporaryRedirect",
-        "TokenCodeInvalidError",
-        "TokenRefreshRequired",
-        "TooManyAccessPoints",
-        "TooManyBuckets",
-        "TooManyMultiRegionAccessPointregionsError",
-        "TooManyMultiRegionAccessPoints",
-        "TooManyTags",
-        "TruncatedInput",
-        "UnauthorizedAccess",
-        "UnauthorizedAccessError",
-        "UnexpectedContent",
-        "UnexpectedIPError",
-        "UnrecognizedFormatException",
-        "UnresolvableGrantByEmailAddress",
-        "UnsupportedArgument",
-        "UnsupportedFunction",
-        "UnsupportedParquetType",
-        "UnsupportedRangeHeader",
-        "UnsupportedScanRangeInput",
-        "UnsupportedSignature",
-        "UnsupportedSqlOperation",
-        "UnsupportedSqlStructure",
-        "UnsupportedStorageClass",
-        "UnsupportedSyntax",
-        "UnsupportedTypeForQuerying",
-        "UserKeyMustBeSpecified",
-        "ValueParseFailure",
-    ];
-
-    #[must_use]
-    fn as_enum_tag(&self) -> usize {
-        match self {
-            Self::AccessControlListNotSupported => 0,
-            Self::AccessDenied => 1,
-            Self::AccessPointAlreadyOwnedByYou => 2,
-            Self::AccountProblem => 3,
-            Self::AllAccessDisabled => 4,
-            Self::AmbiguousFieldName => 5,
-            Self::AmbiguousGrantByEmailAddress => 6,
-            Self::AuthorizationHeaderMalformed => 7,
-            Self::AuthorizationQueryParametersError => 8,
-            Self::BadDigest => 9,
-            Self::BucketAlreadyExists => 10,
-            Self::BucketAlreadyOwnedByYou => 11,
-            Self::BucketHasAccessPointsAttached => 12,
-            Self::BucketNotEmpty => 13,
-            Self::Busy => 14,
-            Self::CSVEscapingRecordDelimiter => 15,
-            Self::CSVParsingError => 16,
-            Self::CSVUnescapedQuote => 17,
-            Self::CastFailed => 18,
-            Self::ClientTokenConflict => 19,
-            Self::ColumnTooLong => 20,
-            Self::ConditionalRequestConflict => 21,
-            Self::ConnectionClosedByRequester => 22,
-            Self::CredentialsNotSupported => 23,
-            Self::CrossLocationLoggingProhibited => 24,
-            Self::DeviceNotActiveError => 25,
-            Self::EmptyRequestBody => 26,
-            Self::EndpointNotFound => 27,
-            Self::EntityTooLarge => 28,
-            Self::EntityTooSmall => 29,
-            Self::EvaluatorBindingDoesNotExist => 30,
-            Self::EvaluatorInvalidArguments => 31,
-            Self::EvaluatorInvalidTimestampFormatPattern => 32,
-            Self::EvaluatorInvalidTimestampFormatPatternSymbol => 33,
-            Self::EvaluatorInvalidTimestampFormatPatternSymbolForParsing => 34,
-            Self::EvaluatorInvalidTimestampFormatPatternToken => 35,
-            Self::EvaluatorLikePatternInvalidEscapeSequence => 36,
-            Self::EvaluatorNegativeLimit => 37,
-            Self::EvaluatorTimestampFormatPatternDuplicateFields => 38,
-            Self::EvaluatorTimestampFormatPatternHourClockAmPmMismatch => 39,
-            Self::EvaluatorUnterminatedTimestampFormatPatternToken => 40,
-            Self::ExpiredToken => 41,
-            Self::ExpressionTooLong => 42,
-            Self::ExternalEvalException => 43,
-            Self::IllegalLocationConstraintException => 44,
-            Self::IllegalSqlFunctionArgument => 45,
-            Self::IllegalVersioningConfigurationException => 46,
-            Self::IncompleteBody => 47,
-            Self::IncorrectEndpoint => 48,
-            Self::IncorrectNumberOfFilesInPostRequest => 49,
-            Self::IncorrectSqlFunctionArgumentType => 50,
-            Self::InlineDataTooLarge => 51,
-            Self::IntegerOverflow => 52,
-            Self::InternalError => 53,
-            Self::InvalidAccessKeyId => 54,
-            Self::InvalidAccessPoint => 55,
-            Self::InvalidAccessPointAliasError => 56,
-            Self::InvalidAddressingHeader => 57,
-            Self::InvalidArgument => 58,
-            Self::InvalidBucketAclWithObjectOwnership => 59,
-            Self::InvalidBucketName => 60,
-            Self::InvalidBucketOwnerAWSAccountID => 61,
-            Self::InvalidBucketState => 62,
-            Self::InvalidCast => 63,
-            Self::InvalidColumnIndex => 64,
-            Self::InvalidCompressionFormat => 65,
-            Self::InvalidDataSource => 66,
-            Self::InvalidDataType => 67,
-            Self::InvalidDigest => 68,
-            Self::InvalidEncryptionAlgorithmError => 69,
-            Self::InvalidExpressionType => 70,
-            Self::InvalidFileHeaderInfo => 71,
-            Self::InvalidHostHeader => 72,
-            Self::InvalidHttpMethod => 73,
-            Self::InvalidJsonType => 74,
-            Self::InvalidKeyPath => 75,
-            Self::InvalidLocationConstraint => 76,
-            Self::InvalidObjectState => 77,
-            Self::InvalidPart => 78,
-            Self::InvalidPartOrder => 79,
-            Self::InvalidPayer => 80,
-            Self::InvalidPolicyDocument => 81,
-            Self::InvalidQuoteFields => 82,
-            Self::InvalidRange => 83,
-            Self::InvalidRegion => 84,
-            Self::InvalidRequest => 85,
-            Self::InvalidRequestParameter => 86,
-            Self::InvalidSOAPRequest => 87,
-            Self::InvalidScanRange => 88,
-            Self::InvalidSecurity => 89,
-            Self::InvalidSessionException => 90,
-            Self::InvalidSignature => 91,
-            Self::InvalidStorageClass => 92,
-            Self::InvalidTableAlias => 93,
-            Self::InvalidTag => 94,
-            Self::InvalidTargetBucketForLogging => 95,
-            Self::InvalidTextEncoding => 96,
-            Self::InvalidToken => 97,
-            Self::InvalidURI => 98,
-            Self::JSONParsingError => 99,
-            Self::KeyTooLongError => 100,
-            Self::LexerInvalidChar => 101,
-            Self::LexerInvalidIONLiteral => 102,
-            Self::LexerInvalidLiteral => 103,
-            Self::LexerInvalidOperator => 104,
-            Self::LikeInvalidInputs => 105,
-            Self::MalformedACLError => 106,
-            Self::MalformedPOSTRequest => 107,
-            Self::MalformedPolicy => 108,
-            Self::MalformedXML => 109,
-            Self::MaxMessageLengthExceeded => 110,
-            Self::MaxOperatorsExceeded => 111,
-            Self::MaxPostPreDataLengthExceededError => 112,
-            Self::MetadataTooLarge => 113,
-            Self::MethodNotAllowed => 114,
-            Self::MissingAttachment => 115,
-            Self::MissingAuthenticationToken => 116,
-            Self::MissingContentLength => 117,
-            Self::MissingRequestBodyError => 118,
-            Self::MissingRequiredParameter => 119,
-            Self::MissingSecurityElement => 120,
-            Self::MissingSecurityHeader => 121,
-            Self::MultipleDataSourcesUnsupported => 122,
-            Self::NoLoggingStatusForKey => 123,
-            Self::NoSuchAccessPoint => 124,
-            Self::NoSuchAsyncRequest => 125,
-            Self::NoSuchBucket => 126,
-            Self::NoSuchBucketPolicy => 127,
-            Self::NoSuchCORSConfiguration => 128,
-            Self::NoSuchKey => 129,
-            Self::NoSuchLifecycleConfiguration => 130,
-            Self::NoSuchMultiRegionAccessPoint => 131,
-            Self::NoSuchObjectLockConfiguration => 132,
-            Self::NoSuchResource => 133,
-            Self::NoSuchTagSet => 134,
-            Self::NoSuchUpload => 135,
-            Self::NoSuchVersion => 136,
-            Self::NoSuchWebsiteConfiguration => 137,
-            Self::NoTransformationDefined => 138,
-            Self::NotDeviceOwnerError => 139,
-            Self::NotImplemented => 140,
-            Self::NotModified => 141,
-            Self::NotSignedUp => 142,
-            Self::NumberFormatError => 143,
-            Self::ObjectLockConfigurationNotFoundError => 144,
-            Self::ObjectSerializationConflict => 145,
-            Self::OperationAborted => 146,
-            Self::OverMaxColumn => 147,
-            Self::OverMaxParquetBlockSize => 148,
-            Self::OverMaxRecordSize => 149,
-            Self::OwnershipControlsNotFoundError => 150,
-            Self::ParquetParsingError => 151,
-            Self::ParquetUnsupportedCompressionCodec => 152,
-            Self::ParseAsteriskIsNotAloneInSelectList => 153,
-            Self::ParseCannotMixSqbAndWildcardInSelectList => 154,
-            Self::ParseCastArity => 155,
-            Self::ParseEmptySelect => 156,
-            Self::ParseExpected2TokenTypes => 157,
-            Self::ParseExpectedArgumentDelimiter => 158,
-            Self::ParseExpectedDatePart => 159,
-            Self::ParseExpectedExpression => 160,
-            Self::ParseExpectedIdentForAlias => 161,
-            Self::ParseExpectedIdentForAt => 162,
-            Self::ParseExpectedIdentForGroupName => 163,
-            Self::ParseExpectedKeyword => 164,
-            Self::ParseExpectedLeftParenAfterCast => 165,
-            Self::ParseExpectedLeftParenBuiltinFunctionCall => 166,
-            Self::ParseExpectedLeftParenValueConstructor => 167,
-            Self::ParseExpectedMember => 168,
-            Self::ParseExpectedNumber => 169,
-            Self::ParseExpectedRightParenBuiltinFunctionCall => 170,
-            Self::ParseExpectedTokenType => 171,
-            Self::ParseExpectedTypeName => 172,
-            Self::ParseExpectedWhenClause => 173,
-            Self::ParseInvalidContextForWildcardInSelectList => 174,
-            Self::ParseInvalidPathComponent => 175,
-            Self::ParseInvalidTypeParam => 176,
-            Self::ParseMalformedJoin => 177,
-            Self::ParseMissingIdentAfterAt => 178,
-            Self::ParseNonUnaryAgregateFunctionCall => 179,
-            Self::ParseSelectMissingFrom => 180,
-            Self::ParseUnExpectedKeyword => 181,
-            Self::ParseUnexpectedOperator => 182,
-            Self::ParseUnexpectedTerm => 183,
-            Self::ParseUnexpectedToken => 184,
-            Self::ParseUnknownOperator => 185,
-            Self::ParseUnsupportedAlias => 186,
-            Self::ParseUnsupportedCallWithStar => 187,
-            Self::ParseUnsupportedCase => 188,
-            Self::ParseUnsupportedCaseClause => 189,
-            Self::ParseUnsupportedLiteralsGroupBy => 190,
-            Self::ParseUnsupportedSelect => 191,
-            Self::ParseUnsupportedSyntax => 192,
-            Self::ParseUnsupportedToken => 193,
-            Self::PermanentRedirect => 194,
-            Self::PermanentRedirectControlError => 195,
-            Self::PreconditionFailed => 196,
-            Self::Redirect => 197,
-            Self::ReplicationConfigurationNotFoundError => 198,
-            Self::RequestHeaderSectionTooLarge => 199,
-            Self::RequestIsNotMultiPartContent => 200,
-            Self::RequestTimeTooSkewed => 201,
-            Self::RequestTimeout => 202,
-            Self::RequestTorrentOfBucketError => 203,
-            Self::ResponseInterrupted => 204,
-            Self::RestoreAlreadyInProgress => 205,
-            Self::ServerSideEncryptionConfigurationNotFoundError => 206,
-            Self::ServiceUnavailable => 207,
-            Self::SignatureDoesNotMatch => 208,
-            Self::SlowDown => 209,
-            Self::TagPolicyException => 210,
-            Self::TemporaryRedirect => 211,
-            Self::TokenCodeInvalidError => 212,
-            Self::TokenRefreshRequired => 213,
-            Self::TooManyAccessPoints => 214,
-            Self::TooManyBuckets => 215,
-            Self::TooManyMultiRegionAccessPointregionsError => 216,
-            Self::TooManyMultiRegionAccessPoints => 217,
-            Self::TooManyTags => 218,
-            Self::TruncatedInput => 219,
-            Self::UnauthorizedAccess => 220,
-            Self::UnauthorizedAccessError => 221,
-            Self::UnexpectedContent => 222,
-            Self::UnexpectedIPError => 223,
-            Self::UnrecognizedFormatException => 224,
-            Self::UnresolvableGrantByEmailAddress => 225,
-            Self::UnsupportedArgument => 226,
-            Self::UnsupportedFunction => 227,
-            Self::UnsupportedParquetType => 228,
-            Self::UnsupportedRangeHeader => 229,
-            Self::UnsupportedScanRangeInput => 230,
-            Self::UnsupportedSignature => 231,
-            Self::UnsupportedSqlOperation => 232,
-            Self::UnsupportedSqlStructure => 233,
-            Self::UnsupportedStorageClass => 234,
-            Self::UnsupportedSyntax => 235,
-            Self::UnsupportedTypeForQuerying => 236,
-            Self::UserKeyMustBeSpecified => 237,
-            Self::ValueParseFailure => 238,
-            Self::Custom(_) => usize::MAX,
-        }
-    }
-
-    pub(crate) fn as_static_str(&self) -> Option<&'static str> {
-        Self::STATIC_CODE_LIST.get(self.as_enum_tag()).copied()
-    }
-
-    #[must_use]
-    pub fn from_bytes(s: &[u8]) -> Option<Self> {
-        let s = std::str::from_utf8(s).ok()?;
-        if let Some(code) = S3_ERROR_CODE_MAP.get(s) {
-            return Some(code.clone());
-        }
-        let lowercase = s.to_ascii_lowercase();
-        if let Some(code) = S3_ERROR_CODE_LOWERCASE_MAP.get(lowercase.as_str()) {
-            return Some(code.clone());
-        }
-        Some(Self::Custom(s.into()))
-    }
-
-    #[allow(clippy::match_same_arms)]
-    #[must_use]
-    pub fn status_code(&self) -> Option<StatusCode> {
-        match self {
-            Self::AccessControlListNotSupported => Some(StatusCode::BAD_REQUEST),
-            Self::AccessDenied => Some(StatusCode::FORBIDDEN),
-            Self::AccessPointAlreadyOwnedByYou => Some(StatusCode::CONFLICT),
-            Self::AccountProblem => Some(StatusCode::FORBIDDEN),
-            Self::AllAccessDisabled => Some(StatusCode::FORBIDDEN),
-            Self::AmbiguousFieldName => Some(StatusCode::BAD_REQUEST),
-            Self::AmbiguousGrantByEmailAddress => Some(StatusCode::BAD_REQUEST),
-            Self::AuthorizationHeaderMalformed => Some(StatusCode::BAD_REQUEST),
-            Self::AuthorizationQueryParametersError => Some(StatusCode::BAD_REQUEST),
-            Self::BadDigest => Some(StatusCode::BAD_REQUEST),
-            Self::BucketAlreadyExists => Some(StatusCode::CONFLICT),
-            Self::BucketAlreadyOwnedByYou => Some(StatusCode::CONFLICT),
-            Self::BucketHasAccessPointsAttached => Some(StatusCode::BAD_REQUEST),
-            Self::BucketNotEmpty => Some(StatusCode::CONFLICT),
-            Self::Busy => Some(StatusCode::SERVICE_UNAVAILABLE),
-            Self::CSVEscapingRecordDelimiter => Some(StatusCode::BAD_REQUEST),
-            Self::CSVParsingError => Some(StatusCode::BAD_REQUEST),
-            Self::CSVUnescapedQuote => Some(StatusCode::BAD_REQUEST),
-            Self::CastFailed => Some(StatusCode::BAD_REQUEST),
-            Self::ClientTokenConflict => Some(StatusCode::CONFLICT),
-            Self::ColumnTooLong => Some(StatusCode::BAD_REQUEST),
-            Self::ConditionalRequestConflict => Some(StatusCode::CONFLICT),
-            Self::ConnectionClosedByRequester => Some(StatusCode::BAD_REQUEST),
-            Self::CredentialsNotSupported => Some(StatusCode::BAD_REQUEST),
-            Self::CrossLocationLoggingProhibited => Some(StatusCode::FORBIDDEN),
-            Self::DeviceNotActiveError => Some(StatusCode::BAD_REQUEST),
-            Self::EmptyRequestBody => Some(StatusCode::BAD_REQUEST),
-            Self::EndpointNotFound => Some(StatusCode::BAD_REQUEST),
-            Self::EntityTooLarge => Some(StatusCode::BAD_REQUEST),
-            Self::EntityTooSmall => Some(StatusCode::BAD_REQUEST),
-            Self::EvaluatorBindingDoesNotExist => Some(StatusCode::BAD_REQUEST),
-            Self::EvaluatorInvalidArguments => Some(StatusCode::BAD_REQUEST),
-            Self::EvaluatorInvalidTimestampFormatPattern => Some(StatusCode::BAD_REQUEST),
-            Self::EvaluatorInvalidTimestampFormatPatternSymbol => Some(StatusCode::BAD_REQUEST),
-            Self::EvaluatorInvalidTimestampFormatPatternSymbolForParsing => Some(StatusCode::BAD_REQUEST),
-            Self::EvaluatorInvalidTimestampFormatPatternToken => Some(StatusCode::BAD_REQUEST),
-            Self::EvaluatorLikePatternInvalidEscapeSequence => Some(StatusCode::BAD_REQUEST),
-            Self::EvaluatorNegativeLimit => Some(StatusCode::BAD_REQUEST),
-            Self::EvaluatorTimestampFormatPatternDuplicateFields => Some(StatusCode::BAD_REQUEST),
-            Self::EvaluatorTimestampFormatPatternHourClockAmPmMismatch => Some(StatusCode::BAD_REQUEST),
-            Self::EvaluatorUnterminatedTimestampFormatPatternToken => Some(StatusCode::BAD_REQUEST),
-            Self::ExpiredToken => Some(StatusCode::BAD_REQUEST),
-            Self::ExpressionTooLong => Some(StatusCode::BAD_REQUEST),
-            Self::ExternalEvalException => Some(StatusCode::BAD_REQUEST),
-            Self::IllegalLocationConstraintException => Some(StatusCode::BAD_REQUEST),
-            Self::IllegalSqlFunctionArgument => Some(StatusCode::BAD_REQUEST),
-            Self::IllegalVersioningConfigurationException => Some(StatusCode::BAD_REQUEST),
-            Self::IncompleteBody => Some(StatusCode::BAD_REQUEST),
-            Self::IncorrectEndpoint => Some(StatusCode::BAD_REQUEST),
-            Self::IncorrectNumberOfFilesInPostRequest => Some(StatusCode::BAD_REQUEST),
-            Self::IncorrectSqlFunctionArgumentType => Some(StatusCode::BAD_REQUEST),
-            Self::InlineDataTooLarge => Some(StatusCode::BAD_REQUEST),
-            Self::IntegerOverflow => Some(StatusCode::BAD_REQUEST),
-            Self::InternalError => Some(StatusCode::INTERNAL_SERVER_ERROR),
-            Self::InvalidAccessKeyId => Some(StatusCode::FORBIDDEN),
-            Self::InvalidAccessPoint => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidAccessPointAliasError => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidAddressingHeader => None,
-            Self::InvalidArgument => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidBucketAclWithObjectOwnership => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidBucketName => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidBucketOwnerAWSAccountID => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidBucketState => Some(StatusCode::CONFLICT),
-            Self::InvalidCast => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidColumnIndex => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidCompressionFormat => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidDataSource => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidDataType => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidDigest => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidEncryptionAlgorithmError => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidExpressionType => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidFileHeaderInfo => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidHostHeader => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidHttpMethod => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidJsonType => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidKeyPath => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidLocationConstraint => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidObjectState => Some(StatusCode::FORBIDDEN),
-            Self::InvalidPart => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidPartOrder => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidPayer => Some(StatusCode::FORBIDDEN),
-            Self::InvalidPolicyDocument => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidQuoteFields => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidRange => Some(StatusCode::RANGE_NOT_SATISFIABLE),
-            Self::InvalidRegion => Some(StatusCode::FORBIDDEN),
-            Self::InvalidRequest => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidRequestParameter => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidSOAPRequest => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidScanRange => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidSecurity => Some(StatusCode::FORBIDDEN),
-            Self::InvalidSessionException => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidSignature => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidStorageClass => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidTableAlias => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidTag => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidTargetBucketForLogging => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidTextEncoding => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidToken => Some(StatusCode::BAD_REQUEST),
-            Self::InvalidURI => Some(StatusCode::BAD_REQUEST),
-            Self::JSONParsingError => Some(StatusCode::BAD_REQUEST),
-            Self::KeyTooLongError => Some(StatusCode::BAD_REQUEST),
-            Self::LexerInvalidChar => Some(StatusCode::BAD_REQUEST),
-            Self::LexerInvalidIONLiteral => Some(StatusCode::BAD_REQUEST),
-            Self::LexerInvalidLiteral => Some(StatusCode::BAD_REQUEST),
-            Self::LexerInvalidOperator => Some(StatusCode::BAD_REQUEST),
-            Self::LikeInvalidInputs => Some(StatusCode::BAD_REQUEST),
-            Self::MalformedACLError => Some(StatusCode::BAD_REQUEST),
-            Self::MalformedPOSTRequest => Some(StatusCode::BAD_REQUEST),
-            Self::MalformedPolicy => Some(StatusCode::BAD_REQUEST),
-            Self::MalformedXML => Some(StatusCode::BAD_REQUEST),
-            Self::MaxMessageLengthExceeded => Some(StatusCode::BAD_REQUEST),
-            Self::MaxOperatorsExceeded => Some(StatusCode::BAD_REQUEST),
-            Self::MaxPostPreDataLengthExceededError => Some(StatusCode::BAD_REQUEST),
-            Self::MetadataTooLarge => Some(StatusCode::BAD_REQUEST),
-            Self::MethodNotAllowed => Some(StatusCode::METHOD_NOT_ALLOWED),
-            Self::MissingAttachment => None,
-            Self::MissingAuthenticationToken => Some(StatusCode::FORBIDDEN),
-            Self::MissingContentLength => Some(StatusCode::LENGTH_REQUIRED),
-            Self::MissingRequestBodyError => Some(StatusCode::BAD_REQUEST),
-            Self::MissingRequiredParameter => Some(StatusCode::BAD_REQUEST),
-            Self::MissingSecurityElement => Some(StatusCode::BAD_REQUEST),
-            Self::MissingSecurityHeader => Some(StatusCode::BAD_REQUEST),
-            Self::MultipleDataSourcesUnsupported => Some(StatusCode::BAD_REQUEST),
-            Self::NoLoggingStatusForKey => Some(StatusCode::BAD_REQUEST),
-            Self::NoSuchAccessPoint => Some(StatusCode::NOT_FOUND),
-            Self::NoSuchAsyncRequest => Some(StatusCode::NOT_FOUND),
-            Self::NoSuchBucket => Some(StatusCode::NOT_FOUND),
-            Self::NoSuchBucketPolicy => Some(StatusCode::NOT_FOUND),
-            Self::NoSuchCORSConfiguration => Some(StatusCode::NOT_FOUND),
-            Self::NoSuchKey => Some(StatusCode::NOT_FOUND),
-            Self::NoSuchLifecycleConfiguration => Some(StatusCode::NOT_FOUND),
-            Self::NoSuchMultiRegionAccessPoint => Some(StatusCode::NOT_FOUND),
-            Self::NoSuchObjectLockConfiguration => Some(StatusCode::NOT_FOUND),
-            Self::NoSuchResource => Some(StatusCode::NOT_FOUND),
-            Self::NoSuchTagSet => Some(StatusCode::NOT_FOUND),
-            Self::NoSuchUpload => Some(StatusCode::NOT_FOUND),
-            Self::NoSuchVersion => Some(StatusCode::NOT_FOUND),
-            Self::NoSuchWebsiteConfiguration => Some(StatusCode::NOT_FOUND),
-            Self::NoTransformationDefined => Some(StatusCode::NOT_FOUND),
-            Self::NotDeviceOwnerError => Some(StatusCode::BAD_REQUEST),
-            Self::NotImplemented => Some(StatusCode::NOT_IMPLEMENTED),
-            Self::NotModified => Some(StatusCode::NOT_MODIFIED),
-            Self::NotSignedUp => Some(StatusCode::FORBIDDEN),
-            Self::NumberFormatError => Some(StatusCode::BAD_REQUEST),
-            Self::ObjectLockConfigurationNotFoundError => Some(StatusCode::NOT_FOUND),
-            Self::ObjectSerializationConflict => Some(StatusCode::BAD_REQUEST),
-            Self::OperationAborted => Some(StatusCode::CONFLICT),
-            Self::OverMaxColumn => Some(StatusCode::BAD_REQUEST),
-            Self::OverMaxParquetBlockSize => Some(StatusCode::BAD_REQUEST),
-            Self::OverMaxRecordSize => Some(StatusCode::BAD_REQUEST),
-            Self::OwnershipControlsNotFoundError => Some(StatusCode::NOT_FOUND),
-            Self::ParquetParsingError => Some(StatusCode::BAD_REQUEST),
-            Self::ParquetUnsupportedCompressionCodec => Some(StatusCode::BAD_REQUEST),
-            Self::ParseAsteriskIsNotAloneInSelectList => Some(StatusCode::BAD_REQUEST),
-            Self::ParseCannotMixSqbAndWildcardInSelectList => Some(StatusCode::BAD_REQUEST),
-            Self::ParseCastArity => Some(StatusCode::BAD_REQUEST),
-            Self::ParseEmptySelect => Some(StatusCode::BAD_REQUEST),
-            Self::ParseExpected2TokenTypes => Some(StatusCode::BAD_REQUEST),
-            Self::ParseExpectedArgumentDelimiter => Some(StatusCode::BAD_REQUEST),
-            Self::ParseExpectedDatePart => Some(StatusCode::BAD_REQUEST),
-            Self::ParseExpectedExpression => Some(StatusCode::BAD_REQUEST),
-            Self::ParseExpectedIdentForAlias => Some(StatusCode::BAD_REQUEST),
-            Self::ParseExpectedIdentForAt => Some(StatusCode::BAD_REQUEST),
-            Self::ParseExpectedIdentForGroupName => Some(StatusCode::BAD_REQUEST),
-            Self::ParseExpectedKeyword => Some(StatusCode::BAD_REQUEST),
-            Self::ParseExpectedLeftParenAfterCast => Some(StatusCode::BAD_REQUEST),
-            Self::ParseExpectedLeftParenBuiltinFunctionCall => Some(StatusCode::BAD_REQUEST),
-            Self::ParseExpectedLeftParenValueConstructor => Some(StatusCode::BAD_REQUEST),
-            Self::ParseExpectedMember => Some(StatusCode::BAD_REQUEST),
-            Self::ParseExpectedNumber => Some(StatusCode::BAD_REQUEST),
-            Self::ParseExpectedRightParenBuiltinFunctionCall => Some(StatusCode::BAD_REQUEST),
-            Self::ParseExpectedTokenType => Some(StatusCode::BAD_REQUEST),
-            Self::ParseExpectedTypeName => Some(StatusCode::BAD_REQUEST),
-            Self::ParseExpectedWhenClause => Some(StatusCode::BAD_REQUEST),
-            Self::ParseInvalidContextForWildcardInSelectList => Some(StatusCode::BAD_REQUEST),
-            Self::ParseInvalidPathComponent => Some(StatusCode::BAD_REQUEST),
-            Self::ParseInvalidTypeParam => Some(StatusCode::BAD_REQUEST),
-            Self::ParseMalformedJoin => Some(StatusCode::BAD_REQUEST),
-            Self::ParseMissingIdentAfterAt => Some(StatusCode::BAD_REQUEST),
-            Self::ParseNonUnaryAgregateFunctionCall => Some(StatusCode::BAD_REQUEST),
-            Self::ParseSelectMissingFrom => Some(StatusCode::BAD_REQUEST),
-            Self::ParseUnExpectedKeyword => Some(StatusCode::BAD_REQUEST),
-            Self::ParseUnexpectedOperator => Some(StatusCode::BAD_REQUEST),
-            Self::ParseUnexpectedTerm => Some(StatusCode::BAD_REQUEST),
-            Self::ParseUnexpectedToken => Some(StatusCode::BAD_REQUEST),
-            Self::ParseUnknownOperator => Some(StatusCode::BAD_REQUEST),
-            Self::ParseUnsupportedAlias => Some(StatusCode::BAD_REQUEST),
-            Self::ParseUnsupportedCallWithStar => Some(StatusCode::BAD_REQUEST),
-            Self::ParseUnsupportedCase => Some(StatusCode::BAD_REQUEST),
-            Self::ParseUnsupportedCaseClause => Some(StatusCode::BAD_REQUEST),
-            Self::ParseUnsupportedLiteralsGroupBy => Some(StatusCode::BAD_REQUEST),
-            Self::ParseUnsupportedSelect => Some(StatusCode::BAD_REQUEST),
-            Self::ParseUnsupportedSyntax => Some(StatusCode::BAD_REQUEST),
-            Self::ParseUnsupportedToken => Some(StatusCode::BAD_REQUEST),
-            Self::PermanentRedirect => Some(StatusCode::MOVED_PERMANENTLY),
-            Self::PermanentRedirectControlError => Some(StatusCode::MOVED_PERMANENTLY),
-            Self::PreconditionFailed => Some(StatusCode::PRECONDITION_FAILED),
-            Self::Redirect => Some(StatusCode::TEMPORARY_REDIRECT),
-            Self::ReplicationConfigurationNotFoundError => Some(StatusCode::NOT_FOUND),
-            Self::RequestHeaderSectionTooLarge => Some(StatusCode::BAD_REQUEST),
-            Self::RequestIsNotMultiPartContent => Some(StatusCode::BAD_REQUEST),
-            Self::RequestTimeTooSkewed => Some(StatusCode::FORBIDDEN),
-            Self::RequestTimeout => Some(StatusCode::BAD_REQUEST),
-            Self::RequestTorrentOfBucketError => Some(StatusCode::BAD_REQUEST),
-            Self::ResponseInterrupted => Some(StatusCode::BAD_REQUEST),
-            Self::RestoreAlreadyInProgress => Some(StatusCode::CONFLICT),
-            Self::ServerSideEncryptionConfigurationNotFoundError => Some(StatusCode::BAD_REQUEST),
-            Self::ServiceUnavailable => Some(StatusCode::SERVICE_UNAVAILABLE),
-            Self::SignatureDoesNotMatch => Some(StatusCode::FORBIDDEN),
-            Self::SlowDown => Some(StatusCode::SERVICE_UNAVAILABLE),
-            Self::TagPolicyException => Some(StatusCode::BAD_REQUEST),
-            Self::TemporaryRedirect => Some(StatusCode::TEMPORARY_REDIRECT),
-            Self::TokenCodeInvalidError => Some(StatusCode::BAD_REQUEST),
-            Self::TokenRefreshRequired => Some(StatusCode::BAD_REQUEST),
-            Self::TooManyAccessPoints => Some(StatusCode::BAD_REQUEST),
-            Self::TooManyBuckets => Some(StatusCode::BAD_REQUEST),
-            Self::TooManyMultiRegionAccessPointregionsError => Some(StatusCode::BAD_REQUEST),
-            Self::TooManyMultiRegionAccessPoints => Some(StatusCode::BAD_REQUEST),
-            Self::TooManyTags => Some(StatusCode::BAD_REQUEST),
-            Self::TruncatedInput => Some(StatusCode::BAD_REQUEST),
-            Self::UnauthorizedAccess => Some(StatusCode::UNAUTHORIZED),
-            Self::UnauthorizedAccessError => Some(StatusCode::FORBIDDEN),
-            Self::UnexpectedContent => Some(StatusCode::BAD_REQUEST),
-            Self::UnexpectedIPError => Some(StatusCode::FORBIDDEN),
-            Self::UnrecognizedFormatException => Some(StatusCode::BAD_REQUEST),
-            Self::UnresolvableGrantByEmailAddress => Some(StatusCode::BAD_REQUEST),
-            Self::UnsupportedArgument => Some(StatusCode::BAD_REQUEST),
-            Self::UnsupportedFunction => Some(StatusCode::BAD_REQUEST),
-            Self::UnsupportedParquetType => Some(StatusCode::BAD_REQUEST),
-            Self::UnsupportedRangeHeader => Some(StatusCode::BAD_REQUEST),
-            Self::UnsupportedScanRangeInput => Some(StatusCode::BAD_REQUEST),
-            Self::UnsupportedSignature => Some(StatusCode::BAD_REQUEST),
-            Self::UnsupportedSqlOperation => Some(StatusCode::BAD_REQUEST),
-            Self::UnsupportedSqlStructure => Some(StatusCode::BAD_REQUEST),
-            Self::UnsupportedStorageClass => Some(StatusCode::BAD_REQUEST),
-            Self::UnsupportedSyntax => Some(StatusCode::BAD_REQUEST),
-            Self::UnsupportedTypeForQuerying => Some(StatusCode::BAD_REQUEST),
-            Self::UserKeyMustBeSpecified => Some(StatusCode::BAD_REQUEST),
-            Self::ValueParseFailure => Some(StatusCode::BAD_REQUEST),
-            Self::Custom(_) => None,
-        }
-    }
-
-    #[must_use]
-    pub fn default_message(&self) -> Option<&'static str> {
-        let s = self.as_static_str()?;
-        S3_ERROR_CODE_DEFAULT_MESSAGE.get(s).copied()
-    }
-}
